+++ mkdocs.yml
@@ -37,9 +37,8 @@ theme:
   # - header.autohide
   icon:
     repo: fontawesome/brands/github
-  # TODO need to add back a logo
-  #logo: assets/img/logo.svg
-  #favicon: assets/img/favicon.png
+  logo: assets/img/dgraph.svg
+  favicon: assets/img/dgraph.png
   language: en
 plugins:
   - search:
