### 2024-04-28 b76e6ba
Added activation example to docs
