+++ docs/examples/celf.md
@@ -1,0 +1,157 @@
+# Influence Maximization
+
+## Motivation and Problem Definition
+
+Suppose we are interested in computing the seed set which activates the most number of nodes in expectation.
+This problem is known as [influence maximization](https://snap-stanford.github.io/cs224w-notes/network-methods/influence-maximization).
+Given a graph $G = (V,E)$ and a fixed budget $k$, we wish to choose a seed set $S \subseteq V$ such that $|S| = k$ and the influence
+$\sigma(S)$ on the graph is maximized.
+
+## Overview of CELF
+
+ A well-known algorithm for this problem is the [CELF](https://hautahi.com/im_greedycelf) algorithm, which is
+an optimized version of the greedy algorithm. At a high level, this algorithm works by greedily adding nodes to
+the seed set which has the highest marginal gain. This is repeated until the budget is exhausted.
+
+## Computing Marginal Gains
+
+To implement this, we first need a specialized function to compute the marginal gain of adding a node to a
+given seed set. `CyNetDiff` allows us to accomplish this task very quickly:
+
+```python
+from cynetdiff.models import DiffusionModel
+
+def compute_marginal_gain(
+    model: DiffusionModel,
+    new_node: int,
+    seeds: set[int],
+    num_trials: int = 1_000,
+) -> float:
+    """
+    Compute the marginal gain in the spread of influence by adding a new node to the set of seed nodes,
+    by summing the differences of spreads for each trial and then taking the average.
+
+    Parameters:
+    - model: The model used for simulating the spread of influence.
+    - new_node: The new node to consider adding to the set of seed nodes.
+    - seeds: The current set of seed nodes.
+    - num_trials: The number of trials to average the spread of influence over.
+
+    Returns:
+    - The average marginal gain in the spread of influence by adding the new node.
+    """
+
+    original_spread = 0
+    new_spread = 0
+    # If no seeds at the beginning, original spread is always just zero.
+    # Prevents wasted work in cases where the seed set is empty.
+    if len(seeds) > 0:
+        model.set_seeds(seeds)
+
+        for _ in range(num_trials):
+            model.reset_model()
+            model.advance_until_completion()
+            original_spread += model.get_num_activated_nodes()
+
+    new_seeds = seeds.union({new_node})
+    model.set_seeds(new_seeds)
+
+    for _ in range(num_trials):
+        model.reset_model()
+        model.advance_until_completion()
+        new_spread += model.get_num_activated_nodes()
+
+    # Avoid floating point division until the very end.
+    return (new_spread - original_spread) / num_trials
+```
+
+## Main Algorithm
+
+With this function, implementing the rest of the algorithm is straightforward.
+
+```python
+from tqdm import tqdm, trange
+from cynetdiff.utils import networkx_to_ic_model
+
+
+def celf(
+    model: DiffusionModel, k: int, num_trials: int = 1_000
+) -> t.Tuple[t.Set[int], t.List[float]]:
+    """
+    Input: graph object, number of seed nodes
+    Output: optimal seed set, resulting spread, time for each iteration
+    Code adapted from this blog post:
+    https://hautahi.com/im_greedycelf
+    """
+
+    # Run the CELF algorithm
+    marg_gain = []
+
+    print("Computing marginal gains.")
+    # First, compute all marginal gains
+    for node in tqdm(list(dir_graph.nodes())):
+        marg_gain.append(
+            (
+                -compute_marginal_gain(
+                    model,
+                    node,
+                    set(),
+                    num_trials,
+                ),
+                node,
+            )
+        )
+
+    # Convert to heap
+    heapq.heapify(marg_gain)
+
[…]
