+++ mkdocs.yml
@@ -83,6 +83,8 @@ markdown_extensions:
         - name: mermaid
           class: mermaid
           format: !!python/name:pymdownx.superfences.fence_code_format
+  - pymdownx.arithmatex:
+      generic: true
   - pymdownx.inlinehilite
   - pymdownx.snippets
   - pymdownx.critic
@@ -104,7 +106,10 @@ extra:
       name: Package on PyPI
 # extra_css:
 #   - stylesheets/extra.css
-# extra_javascript:
+extra_javascript:
+  - javascripts/mathjax.js
+  - https://polyfill.io/v3/polyfill.min.js?features=es6
+  - https://cdn.jsdelivr.net/npm/mathjax@3/es5/tex-mml-chtml.js
 #   - javascripts/extra.js
 copyright: Copyright &copy; 2024 Eliot Robson, Abhishek K. Umrawal
 nav:
