### 2024-04-29 1b9d13d
Refactored LT model for testing purposes
