+++ src/cynetdiff/models.cpp
@@ -1549,11 +1549,11 @@ static const char *__pyx_filename;
 /* #### Code section: filename_table ### */
 
 static const char *__pyx_f[] = {
-  "src/cynetdiff/models.pyx",
+  "src\\\\cynetdiff\\\\models.pyx",
   "<stringsource>",
   "contextvars.pxd",
   "array.pxd",
-  "src/cynetdiff/models.pxd",
+  "src\\\\cynetdiff\\\\models.pxd",
   "type.pxd",
   "bool.pxd",
   "complex.pxd",
@@ -1830,8 +1830,8 @@ struct __pyx_obj_9cynetdiff_6models___pyx_scope_struct_1_get_activated_nodes {
 };
 
 
-/* "cynetdiff/models.pyx":221
- *         self.buckets.clear()
+/* "cynetdiff/models.pyx":216
+ *         self.thresholds.clear()
  * 
  *     def get_newly_activated_nodes(self):             # <<<<<<<<<<<<<<
  *         for node in self.work_deque:
@@ -1845,7 +1845,7 @@ struct __pyx_obj_9cynetdiff_6models___pyx_scope_struct_2_get_newly_activated_nod
 };
 
 
-/* "cynetdiff/models.pyx":225
+/* "cynetdiff/models.pyx":220
  *             yield node
  * 
  *     def get_activated_nodes(self):             # <<<<<<<<<<<<<<
@@ -1980,7 +1980,7 @@ static int __pyx_f_9cynetdiff_6models_23IndependentCascadeModel__IndependentCasc
 
 struct __pyx_vtabstruct_9cynetdiff_6models_LinearThresholdModel {
   struct __pyx_vtabstruct_9cynetdiff_6models_DiffusionModel __pyx_base;
-  void (*reassign_thresholds)(struct __pyx_obj_9cynetdiff_6models_LinearThresholdModel *, int __pyx_skip_dispatch);
+  void (*_assign_thresholds)(struct __pyx_obj_9cynetdiff_6models_LinearThresholdModel *, arrayobject *, int __pyx_skip_dispatch);
   int (*_LinearThresholdModel__advance_model)(struct __pyx_obj_9cynetdiff_6models_LinearThresholdModel *, std::deque<unsigned int>  &, std::unordered_set<unsigned int>  &, std::unordered_map<unsigned int,float>  &, std::unordered_map<unsigned int,float>  &);
 };
 static struct __pyx_vtabstruct_9cynetdiff_6models_LinearThresholdModel *__pyx_vtabptr_9cynetdiff_6models_LinearThresholdModel;
@@ -3323,7 +3323,7 @@ static int __pyx_f_9cynetdiff_6models_23IndependentCascadeModel__IndependentCasc
 static void __pyx_f_9cynetdiff_6models_23IndependentCascadeModel_advance_until_completion(struct __pyx_obj_9cynetdiff_6models_IndependentCascadeModel *__pyx_v_self, int __pyx_skip_dispatch); /* proto*/
 static void __pyx_f_9cynetdiff_6models_23IndependentCascadeModel_advance_model(struct __pyx_obj_9cynetdiff_6models_IndependentCascadeModel *__pyx_v_self, int __pyx_skip_dispatch); /* proto*/
[…]
