+++ src/cynetdiff/models.cpp
@@ -1971,7 +1971,7 @@ static int __pyx_f_9cynetdiff_6models_23IndependentCascadeModel__IndependentCasc
 
 
 /* "cynetdiff/models.pyx":150
- * 
+ * from libc.stdio cimport printf
  * # LT Model
  * cdef class LinearThresholdModel(DiffusionModel):             # <<<<<<<<<<<<<<
  *     # Functions that interface with the Python side of things
@@ -23008,7 +23008,7 @@ static int __pyx_f_9cynetdiff_6models_23IndependentCascadeModel__IndependentCasc
  *                     work_deque.push_back(child)
  *                     seen_set.insert(child)             # <<<<<<<<<<<<<<
  * 
- * 
+ * from libc.stdio cimport printf
  */
         try {
           __pyx_v_seen_set.insert(__pyx_v_child);
@@ -25795,7 +25795,7 @@ static int __pyx_f_9cynetdiff_6models_20LinearThresholdModel__LinearThresholdMod
  * 
  *                     if thresholds.count(child) == 0:             # <<<<<<<<<<<<<<
  *                         thresholds[child] = next_rand()
- * 
+ *                         while thresholds[child] == 0.0:
  */
         __pyx_t_5 = (__pyx_v_thresholds.count(__pyx_v_child) == 0);
         if (__pyx_t_5) {
@@ -25804,23 +25804,45 @@ static int __pyx_f_9cynetdiff_6models_20LinearThresholdModel__LinearThresholdMod
  * 
  *                     if thresholds.count(child) == 0:
  *                         thresholds[child] = next_rand()             # <<<<<<<<<<<<<<
- * 
- *                     threshold = thresholds[child]
+ *                         while thresholds[child] == 0.0:
+ *                             thresholds[child] = next_rand()
  */
           __pyx_t_13 = __pyx_f_9cynetdiff_6models_next_rand(); if (unlikely(__pyx_t_13 == ((double)-1) && __Pyx_ErrOccurredWithGIL())) __PYX_ERR(0, 275, __pyx_L1_error)
           (__pyx_v_thresholds[__pyx_v_child]) = __pyx_t_13;
+
+          /* "cynetdiff/models.pyx":276
+ *                     if thresholds.count(child) == 0:
+ *                         thresholds[child] = next_rand()
+ *                         while thresholds[child] == 0.0:             # <<<<<<<<<<<<<<
+ *                             thresholds[child] = next_rand()
+ * 
+ */
+          while (1) {
+            __pyx_t_5 = ((__pyx_v_thresholds[__pyx_v_child]) == 0.0);
+            if (!__pyx_t_5) break;
+
+            /* "cynetdiff/models.pyx":277
+ *                         thresholds[child] = next_rand()
+ *                         while thresholds[child] == 0.0:
+ *                             thresholds[child] = next_rand()             # <<<<<<<<<<<<<<
+ * 
+ *                     threshold = thresholds[child]
+ */
+            __pyx_t_13 = __pyx_f_9cynetdiff_6models_next_rand(); if (unlikely(__pyx_t_13 == ((double)-1) && __Pyx_ErrOccurredWithGIL())) __PYX_ERR(0, 277, __pyx_L1_error)
+            (__pyx_v_thresholds[__pyx_v_child]) = __pyx_t_13;
+          }
 
           /* "cynetdiff/models.pyx":274
  *                         buckets[child] = 0.0
  * 
  *                     if thresholds.count(child) == 0:             # <<<<<<<<<<<<<<
  *                         thresholds[child] = next_rand()
- * 
[…]
