### 2024-04-29 0a10e8a
Update activated.md
