### 2024-04-29 c1651d2
docs: Added examples to docs and bumped version (#13)
Minor changes to make example code in docs cleaner.
