+++ docs/contributing.md
@@ -1,3 +1,1 @@
 --8<-- "CONTRIBUTING.md"
-
-<!-- TODO figure out how to get link to code of conduct working ok. -->
