+++ docs/people.md
@@ -1,7 +1,8 @@
 ## Maintainers
 
 - [eliotwrobson](https://github.com/eliotwrobson)
+- [abhishekumrawal](https://github.com/abhishekumrawal)
 
 ## Contributors
 
-- [abhishekumrawal](https://github.com/abhishekumrawal)
+- Dhemath Reddy
