### 2024-04-29 #175 @eliotwrobson
@Batalex I have just published a release with the following changes:
- Updated the elevator pitch and example in the README
- Added two longer extended examples to the documentation site
