### 2024-04-30 c62e6fc
Added more tests for utility function
