+++ docs/examples/preliminaries.md
@@ -57,10 +57,10 @@ obtaining useful information from these models.
 ## Independent Cascade Model
 
 In the independent cascade model, when a node $v \in V$ first becomes active, it will be given a single chance
-to activate each currently inactive neighbor $w$. The activation succeeds with probability $p_{v,w}$
-(independent of the history thus far). If $w$ has multiple newly activated neighbors, their attempts occur in
-an arbitrary order. If $v$ succeeds, then $w$ will become active; but whether or not $v$ succeeds, it cannot make
-any further attempts to activate $w$ in subsequent rounds.
+to activate each currently inactive neighbor $u$. The activation succeeds with probability $p_{v,u}$
+(independent of the history thus far). If $u$ has multiple newly activated neighbors, their attempts occur in
+an arbitrary order. If $v$ succeeds, then $u$ will become active; but whether or not $v$ succeeds, it cannot make
+any further attempts to activate $u$ in subsequent rounds.
 
 `CyNetDiff` supports simulating the above process by the use of a model class. This class stores the state
 of the diffusion model, and can be manually advanced by repeated function calls. To start, if we have a
@@ -71,16 +71,16 @@ from cynetdiff.utils import networkx_to_ic_model
 model = networkx_to_ic_model(graph)
 ```
 
-In the above, the activation probabilities are all set to a default of `0.1`. There are a number of different
+By default, the activation probabilities are all set to a default of `0.1`. There are a number of different
 commonly-used weighting schemes which can be set by `CyNetDiff`.
 
 ### Weighting Schemes
 
 In the IC model, the weighting scheme is how the probabilities $p_{v,w}$ are chosen. There are three common weighting schemes handled by utility functions from `CyNetDiff`:
 
-- We can choose the weights $p_{v,w} \in [0,1]$, uniformly at random. This can be set by calling the function `set_activation_uniformly_random(graph)` on a given graph.
-- We can choose the weights $p_{v,w}$, uniformly at random from a set `weight_set`. This can be set by calling the function `set_activation_random_sample(graph, weight_set)` on a given graph.
-- We can choose the weights $p_{v,w} = \frac{1}{|N^-(w)|}$ as the inverse of the in-degree of $w$. This can be set by calling the function `set_activation_weighted_cascade(graph)` on a given graph. The input graph must be _directed_ to use this weighting scheme.
+- We can choose the weights $p_{v,u} \in [0,1]$, uniformly at random. This can be set by calling the function `set_activation_uniformly_random(graph)`.
+- We can choose the weights $p_{v,u}$, uniformly at random from a set `weight_set`. This can be set by calling the function `set_activation_random_sample(graph, weight_set)`.
+- We can choose the weights $p_{v,u} = \frac{1}{|N^{in}(u)|}$ as the inverse of the in-degree of $u$. This can be set by calling the function `set_activation_weighted_cascade(graph)`. The input graph must be _directed_ to use this weighting scheme.
 
 To apply a given weighting scheme, we simply call the corresponding helper function from the utility module
 of `CyNetDiff`. For example, to use the well-known trivalency weighting scheme, simply call:
@@ -96,3 +96,23 @@ model = networkx_to_ic_model(small_world_graph)
 Custom weights can be set under the `"activation_prob"` data key on each edge in the starting `NetworkX` graph.
 
 ## Linear Threshold Model
+
+In the linear threshold model, each directed edge $(v,u)$ is given an influence $w_{v,u} \in [0,1]$ such that
+$\sum_{v \in N^{in}(u)} w_{v,u} \leq 1$, so that the total incoming influence at every node is at
[…]
