+++ README.md
@@ -10,7 +10,9 @@
 
 - **Documentation**: https://eliotwrobson.github.io/CyNetDiff/
 
-Network diffusion processes aim to model the spread of trends through social networks, represented using graphs. Experimental work with these models usually involves simulating these processes many times over large graphs, which can be computationally very expensive. To address this, CyNetDiff is a Cython module implementing the independent cascade and linear threshold models. Development has been focused on performance, while still giving an intuitive, high-level interface to assist in research tasks. To learn more about these specific models, read [this](https://www.researchgate.net/publication/300470631_The_Independent_Cascade_and_Linear_Threshold_Models).
+Network diffusion processes aim to model the spread of trends through social networks, represented using graphs. Experimental work with these models usually involves simulating these processes many times over large graphs, which can be computationally very expensive. To address this, CyNetDiff is a Cython module implementing the independent cascade and linear threshold models. Development has been focused on performance, while still giving an intuitive, high-level interface to assist in research tasks. To learn more about these specific models, read
+[this book chapter](https://www.researchgate.net/publication/300470631_The_Independent_Cascade_and_Linear_Threshold_Models)
+or the [preliminaries page in the documentation](https://eliotwrobson.github.io/CyNetDiff/examples/preliminaries/).
 
 ## Quick Start
 
