### 2024-05-01 a1abfcc
feat: Add various useful revisions (#16)
Expanded documentation and added CSR graph checu function. Also, returned node mapping from utility functions.
