### 2024-05-10 #175 @eliotwrobson
@Batalex I know we don't have an editor yet, but is it too early to begin suggesting reviewers? We've already found someone that might be a good candidate for this.
