### 2024-06-29 #175 @Batalex
Hello @eliotwrobson,
Unfortunately, I do not have an editor at the moment. There are two reasons for that: first, we are so successful our current editors have their hands full currently ([let me brag a little](https://www.pyopensci.org/peer-review-metrics/pyopensci-peer-review-stats#summary-number-of-issues-by-year)). Second, the domain is more niche than others, so we (or more specifically I) cannot rely on personal connections or previous involvements to find someone with a relevant expertise. What would be your target conference? I'll see what I can do to help accelerate the process.
