### 2024-06-30 #175 @sneakers-the-rat
I can edit this, although I just was the editor for the `automata` package and maybe the authors want a break from me ;)
