+++ examples/demo_reqirements.txt
@@ -1,0 +1,9 @@
+ndlib==5.1.1
+networkx==3.2.1
+tqdm==4.66.2
+pooch==1.8.1
+coloraide==3.1.2
+cynetdiff==0.1.13
+matplotlib==3.9.2
+ipywidgets==8.1.5
+pygraphviz==1.13
