### 2024-10-01 #175 @lwasser
can we help @eliotwrobson @sneakers-the-rat with the second reviewer? if you have one domain-specific review, we can pull a second from our signup who knows about packaging and can test for usability. i just updated the label to seeking-reviewers as well. please let me know how we can move things forward. I want to support everyone here!
