### 2024-10-01 #175 @eliotwrobson
@lwasser your assessment is correct, the first reviewer we have is knowledgeable about the domain, so I think having a reviewer focused on packaging / usability / documentation is the way to go 👍🏽
