### 2024-10-14 eliotwrobson/CyNetDiff#21 @eliotwrobson
Adopt SPEC-7 Standard
See title and docs here: https://scientific-python.org/specs/spec-0007/

From review discussion here: https://github.com/pyOpenSci/software-submission/issues/175#issuecomment-2439393615
