### 2024-10-14 #175 @lwasser
hey @eliotwrobson @sneakers-the-rat  I think I found a second reviewer for this @Kai-Striega . 🚀  Kai if you are still able to review this package with a focus on the Cython side of things, we'd appreciate it!
