### 2024-10-18 #175 @lwasser
ok  let's kick off this review!
Below is the editor template that we need to get started with.

## Editor response to review:

---

## Editor comments
:wave: Hi @Kai-Striega @jeremyzhangsq ! Thank you for volunteering to review
for pyOpenSci! It's so awesome to have you here supporting our open source / open science community 🚀

To begin there are a few items to complete:

Before beginning your review, can you both please fill out our [pre-review survey](https://forms.gle/F9mou7S3jhe8DMJ16). This helps us improve all aspects of our review and better understand our community. No personal data will be shared from this survey--it will only be used in an aggregated format by our Executive Director to improve our processes and programs.
- [ ]  @Kai-Striega survey completed.
- [ ] @jeremyzhangsq survey completed.

Thank you in advance for doing this!

## Review templates

here as a comment once your review is complete.

Let's shoot for a review deadline of 8 November 2024.  Please get in touch with any questions that you have!!  Have an awesome weekend y'all!! 🚀
---

Reviewers:
Due date: 8 November 2024.
