### 2024-10-20 #175 @Kai-Striega
A quick couple of questions @lwasser/@sneakers-the-rat:

* as previously discussed, my review has focused primarily on the _Cython_ aspects of the project. Do you require me to review the remaining points on the review template too? I'm happy to do this, but I need to be aware that it's required.
* I've look through the package and written plenty of notes but haven't typed them into the review template yet. As this is my first time reviewing, would we be able to organise a mentoring session to ensure I do a great job?
