### 2024-10-26 #175 @Kai-Striega
@eliotwrobson please see my review below. I'm happy to expand/talk about any of the points I have made.

@sneakers-the-rat / @lwasser I haven't received a mentor yet, but had the available time to do it today. Please let me know if I am required to change/update anything.

  - Instructions provided, however I had difficulty following them. See comments below.
  - Vignette in README and examples/ directory
  - No examples for user facing documentation page
  - No examples in ``models.pyi`` file with rest of documentation

      - PyPi supported and badged
      - conda not yet available, however this package seems appropriate for conda

  - Link to examples/ directory missing in README
  - Vignettes/examples are sufficient
  - Not provided, have not done a lit review to check if this is required.

  - Functions/methods have documentation, but not examples in the documentation.
  - See comments in reviewer comments about wheels and required C++ compiler

  - No performance claims made
  - Benchmarks appear fairly ad-hoc.
  - Benchmarks provided as part of the examples, should be own directory.
  - [asv](https://github.com/airspeed-velocity/asv) (or another benchmarking framework) could be used to produce more reliable benchmarks
      - Missing CPython 3.13 support, however GitHub declares ``Requires: Python >=3.9 ``
      - Missing free-threaded Python support, this is not unusual as the project heavily uses Cython, which does not officially support free-threaded Python yet
      - Passes with ``ruff`` locally
      - Passes in CI for ``ruff`` and ``mypy``
      - As this project utilizes Cython heavily a tool such as [cython-lint](https://github.com/MarcoGorelli/cython-lint) is recommended

---

I have several review comments below. These are ordered by importance.

##### Requires a C++ compiler to build/lack of wheels

Wheels are provided only CPython 3.11 on Linux. This requires all other possible installers to compile the project from source. This is fairly simple, however it provides an extra step/requirements which is possible to avoid. My recommendation is to provide wheels for all officially supported Python versions on all major platforms.

##### Cython code generation

The project currently caches the file ``src/cynetdiff/models.cpp`` explicitly. I usually recommended not to ship the ``.c/.cpp`` files directly. There are several problems with shipping ``.c/.cpp`` files:

* They are not guaranteed to be forward compatible with future CPython C-API or Cython versions. The CPython C-API seems to be changing fairly frequently at the moment.
* It's possible that Cython generated code will not work on certain platforms e.g. arm if built for x86

This is echoed by the official Cython [user guide](https://github.com/cython/cython/blob/master/docs/src/userguide/source_files_and_compilation.rst#distributing-cython-modules)

>  it is not recommended to include generated files in source distributions. Instead, require Cython at build-time to generate the C/C++ files

##### Pseudo random number generation

The package currently uses C's ``rand`` function. From ``models.pyx``:

```cython
from libc.stdlib cimport rand, RAND_MAX
cdef double RAND_SCALE = 1.0 / RAND_MAX

cdef inline double next_rand() nogil:
    return rand() * RAND_SCALE
```

There are several issues with ``rand``:

1. There are no guarantees about the algorithm, and therefore the quality of the randomness, used to generate random numbers
2. It uses global state which can lead to problems with multithreaded applications. While the current project doesn't use any parallelism, this could be significant if [CyNetDiff-1](https://github.com/eliotwrobson/CyNetDiff/issues/1) is implemented.

My recommendation is to adopt a more suitable random number generation engine. In the Cython world there are two ways I can recommend of doing this:

1. C++'s ``<random>`` header via the Cython [bindings](https://github.com/cython/cython/blob/master/Cython/Includes/libcpp/random.pxd)
2. Adopt NumPy's random number generation. This is officially [documented](https://numpy.org/doc/stable/reference/random/extending.html#cython) and an example can be seen [here](https://gist.github.com/ev-br/3d3e5b9682ef8c147e457e9cd2b190a0)

Note that (2) would also allow you to be compliant with [SPEC-7](https://scientific-python.org/specs/spec-0007/).

##### Minor Cython related improvements

Note these points are minor and should not be considered blocking.

######  Use of ``array.array``

The package uses ``array.array`` both for type hints and Cython typing in several places. Consider using Cython's [memoryview](https://cython.readthedocs.io/en/latest/src/userguide/memoryviews.html) instead. memoryviews provide the advantage of being available to anything that provides the [Buffer Protocol](https://docs.python.org/3/c-api/buffer.html), such as NumPy arrays, which I suspect are used significantly more frequently than ``array.array``s are by the packages target audience.

###### Missing potential bounds checks when running cythonize
[1/1] Cythonizing src/cynetdiff/models.pyx
warning: src/cynetdiff/models.pyx:97:51: Use boundscheck(False) for faster access
warning: src/cynetdiff/models.pyx:107:35: Use boundscheck(False) for faster access
warning: src/cynetdiff/models.pyx:135:39: Use boundscheck(False) for faster access
warning: src/cynetdiff/models.pyx:137:38: Use boundscheck(False) for faster access
warning: src/cynetdiff/models.pyx:141:34: Use boundscheck(False) for faster access
warning: src/cynetdiff/models.pyx:261:39: Use boundscheck(False) for faster access
warning: src/cynetdiff/models.pyx:263:45: Use boundscheck(False) for faster access
warning: src/cynetdiff/models.pyx:264:34: Use boundscheck(False) for faster access
warning: src/cynetdiff/models.pyx:268:38: Use boundscheck(False) for faster access
warning: src/cynetdiff/models.pyx:284:52: Use boundscheck(False) for faster access
```
* It may be possible to improve performance by disabling bounds checking in these situations.
* I have not verified if this actually improves performance in these particular circumstances

###### Review resizing ``array.array``
```cython
    for i in range(m):
        influence_arr.append(1.0 / in_degrees[self.edges[i]])
```
This computations resizes ``influence_arr`` multiple times, leading to multiple memory allocations. It would be more efficient to use ``.extend`` rather than ``.append``.

###### Use of name mangling

Starting a name with ``__`` introduces name mangling in Python. In C/C++ these are reserved for the compilers internal keywords. Name mangling can make life difficult. CyNetDiff uses name mangling at several points (e.g. ``__activation_suceeds`` and ``__advance_model``). I'd be careful with these.
