### 2024-10-27 #175 @eliotwrobson
@Kai-Striega thank you for your detailed comments! I have a couple of questions, but I'll try to get started addressing some of these items soon.

> Requires a C++ compiler to build/lack of wheels

Do you know of any documentation that has discussion of how to do this? I agree it is much more convenient to not need a compiler installed, but I wasn't sure how to set this up.

> Cython code generation

This makes sense to me, I could swear I read the opposite somewhere but I must have been mistaken.

> Pseudo random number generation

Thank you for the references! I had actually thought about whether there was a standard way to handle the random number generation. I think I will follow option (2) and add the SPEC-7 compliance.

> Use of array.array

Is there a recommended type annotation that expresses what works with the Cython memoryview? It would be nice to add this annotation along with support for numpy arrays.
