### 2024-10-27 #175 @Kai-Striega
> @Kai-Striega thank you for your detailed comments! I have a couple of questions, but I'll try to get started addressing some of these items soon.

You're welcome.

Feel free to message if you have any more questions. I'll try answer your questions below.

> Do you know of any documentation that has discussion of how to do this? I agree it is much more convenient to not need a compiler installed, but I wasn't sure how to set this up.

AFAIK, there isn't much. You could look at how [SciPy](https://github.com/scipy/scipy/blob/main/.github/workflows/wheels.yml) builds their wheels for nightly builds. This should include all the info you need. I'm happy to help out with this, but might be a bit slow at the moment.

Note that SciPy's CI is far more complex than I would expect for this project.

> This makes sense to me, I could swear I read the opposite somewhere but I must have been mistaken.

This isn't cut and dry. In the past, it was advised to do the opposite. This has changed over the last decade or so. Some people still prefer to include the `.c/cpp` files. However, this isn't recommended anymore.

> Thank you for the references! I had actually thought about whether there was a standard way to handle the random number generation. I think I will follow option (2) and add the SPEC-7 compliance.

You're welcome :)

> Is there a recommended type annotation that expresses what works with the Cython memoryview? It would be nice to add this annotation along with support for numpy arrays.

Unfortunately, not that I've come across. The problem is that Cython memoryviews support anything that supports the [Buffer Protocol](https://docs.python.org/3/c-api/buffer.html) this could include custom objects.

My usual tactic is to use [ArrayLike](https://numpy.org/doc/stable/reference/typing.html#numpy.typing.ArrayLike) with a Python level wrapper, check if the value is scalar and then convert it to an array. This isn't great.

An alternative option would be to create your own [Protocol](https://typing.readthedocs.io/en/latest/spec/protocol.html) with all the buffer protocol methods.
