### 2024-10-27 bad2846
Add Cython lint
