+++ .github/workflows/tests.yml
@@ -33,10 +33,5 @@ jobs:
       - name: Install dependencies
         run: poetry install --with dev
 
-      #- name: Install dependencies
-      #  run: |
-      #    python -m pip install -r build_requirements.txt
-      #    python -m pip install -r test_requirements.txt
-      #    python setup.py build_ext --inplace
       - name: Test with pytest
         run: poetry run pytest
