### 2024-10-28 #175 @Kai-Striega
I'm at work right now, and can't respond in the length I'd like. I'll edit this over the next few days to expand on my thoughts.

TLDR: Neither. SPEC-7 allows for a couple of different options.

Here is the officially recommended way

```python
from collections.abc import Sequence
import numpy as np

SeedLike = int | np.integer | Sequence[int] | np.random.SeedSequence
RNGLike = np.random.Generator | np.random.BitGenerator

def my_func(*, rng: RNGLike | SeedLike | None = None):
    """My function summary.

    Parameters
    ----------
    rng : `numpy.random.Generator`, optional
        Pseudorandom number generator state. When `rng` is None, a new
        `numpy.random.Generator` is created using entropy from the
        operating system. Types other than `numpy.random.Generator` are
        passed to `numpy.random.default_rng` to instantiate a `Generator`.
    """
    rng = np.random.default_rng(rng)

    ...
```

I would use this as a wrapper for your Cython functions.
