### 2024-10-28 a6ed28e
First version of marginal gain function.
