### 2024-10-28 c5a9404
Add marginal gain tests
