### 2024-10-29 1db65d0
Got first part of IC model changes working
