+++ tests/test_lt_model.py
@@ -151,13 +151,22 @@ def _influence_sum(G, froms, to):
 
 
 def generate_random_graph_from_seed(
-    n: int, p: float, directed: bool, include_influence: bool, seed: int = 12345
+    n: int,
+    p: float,
+    directed: bool,
+    include_influence: bool,
+    include_payoff: bool,
+    *,
+    seed: int = 12345,
 ) -> nx.DiGraph:
     graph = nx.fast_gnp_random_graph(n, p, seed=seed, directed=directed).to_directed()
 
     random.seed(seed)
     for _, data in graph.nodes(data=True):
         data["threshold"] = random.random()
+
+        if include_payoff:
+            data["payoff"] = random.uniform(1.0, 10.0)
 
     if include_influence:
         for _, _, data in graph.edges(data=True):
@@ -199,7 +208,11 @@ def test_specific_model(directed: bool, nondefault_influence: bool) -> None:
     p = 0.05
     k = 10
     test_graph = generate_random_graph_from_seed(
-        n, p, directed=directed, include_influence=nondefault_influence
+        n,
+        p,
+        directed=directed,
+        include_influence=nondefault_influence,
+        include_payoff=False,
     )
 
     nodes = list(test_graph.nodes)
@@ -239,7 +252,13 @@ def test_specific_model(directed: bool, nondefault_influence: bool) -> None:
 def test_invalid_seed_error() -> None:
     n = 1000
     p = 0.01
-    test_graph = generate_random_graph_from_seed(n, p, False, True)
+    test_graph = generate_random_graph_from_seed(
+        n,
+        p,
+        False,
+        True,
+        False,
+    )
 
     model, _ = networkx_to_lt_model(test_graph)
 
@@ -254,10 +273,27 @@ def test_invalid_seed_error() -> None:
         model.advance_until_completion()
 
 
+def compute_graph_marginal_gain(
+    graph: nx.DiGraph,
+    new_set_lists: t.List[t.List[int]],
+) -> float:
+    result = 0.0
+
+    new_set = set(item for sublist in new_set_lists for item in sublist)
+
+    for elem in new_set:
+        result += graph.nodes[elem].get("payoff", 1.0)
+
+    return result
+
+
 @pytest.mark.parametrize("directed", [True, False])
 @pytest.mark.parametrize("nondefault_influence", [True, False])
+@pytest.mark.parametrize("include_payoffs", [True, False])
 @pytest.mark.parametrize("seed", [12345, 505050])
-def test_marginal_gain(directed: bool, nondefault_influence: bool, seed: int) -> None:
+def test_marginal_gain(
+    directed: bool, nondefault_influence: bool, include_payoffs: bool, seed: int
+) -> None:
     """
     Test the marginal gain function under a couple of parameter settings.
     """
@@ -266,7 +302,7 @@ def test_marginal_gain(directed: bool, nondefault_influence: bool, seed: int) ->
     p = 0.01
     k = 10
     test_graph = generate_random_graph_from_seed(
-        n, p, directed, nondefault_influence, seed
+        n, p, directed, nondefault_influence, include_payoffs, seed=seed
     )
 
     nodes = list(test_graph.nodes)
@@ -279,30 +315,28 @@ def test_marginal_gain(directed: bool, nondefault_influence: bool, seed: int) ->
     result = model.compute_marginal_gain(
         seeds, None, 1, _node_thresholds=node_thresholds
     )
-    total_activated = float(
-        sum(len(level) for level in linear_threshold(test_graph, seeds))
+    total_activated = compute_graph_marginal_gain(
+        test_graph, linear_threshold(test_graph, seeds)
     )
 
-    assert math.isclose(result, total_activated)
+    assert math.isclose(result, total_activated, abs_tol=0.05)
 
[…]
