+++ src/cynetdiff/utils.py
@@ -196,6 +196,9 @@ def networkx_to_lt_model(
     Converts a NetworkX graph into a Linear Threshold model. Includes influence
     values if they are defined on each edge under the key `"influence"`.
 
+    Payoffs for each node are included if they are defined on each node under the key
+    `"payoff"`.
+
     Parameters
     ----------
     graph : nx.Graph or nx.DiGraph
@@ -209,21 +212,29 @@ def networkx_to_lt_model(
         the model.
     """
 
-    node_list = list(enumerate(graph.nodes()))
-    node_mapping = {node: i for i, node in node_list}
+    node_list = list(enumerate(graph.nodes(data=True)))
+    node_mapping = {node: i for i, (node, _) in node_list}
 
     starts = array.array("I")
     edges = array.array("I")
 
     influence = None
+    payoffs = None
+
+    if next(iter(graph.nodes.data("payoff", None)))[1] is not None:
+        payoffs = array.array("f")
 
     if next(iter(graph.edges.data("influence", None)))[2] is not None:
         influence = array.array("f")
 
     curr_successor = 0
-    for _, node in node_list:
+    for _, (node, data_dict) in node_list:
         # First, add to out neighbors
         starts.append(curr_successor)
+
+        if payoffs is not None:
+            payoffs.append(data_dict["payoff"])
+
         for successor in graph.successors(node):
             other = node_mapping[successor]
             curr_successor += 1
@@ -253,6 +264,7 @@ def networkx_to_lt_model(
     model = LinearThresholdModel(
         starts,
         edges,
+        payoffs=payoffs,
         influence=influence,
     )
 
