### 2024-10-29 9a716bf
Responses to review comments (#22)
Add initial changes responding to review comments and implementing new features. Ready for release.
