+++ src/cynetdiff/models.pyx
@@ -6,8 +6,7 @@ from libcpp.algorithm cimport fill
 from libcpp.vector cimport vector as cvector
 from libcpp.unordered_map cimport unordered_map as cmap
 
-import array
-import random
+cimport cython
 
 # Next, utility functions
 # TODO move these to a separate file later
@@ -38,24 +37,44 @@ cdef class DiffusionModel:
     cpdef void advance_until_completion(self):
         raise NotImplementedError
 
+    cdef float _compute_payoff(
+        self,
+        cset[unsigned int]& new_set,
+        cset[unsigned int]& old_set,
+        float[:] payoffs,
+    ):
+        cdef float result = 0.0
+
+        if payoffs is not None:
+            for node in new_set:
+                if old_set.find(node) == old_set.end():
+                    result += payoffs[node]
+        else:
+            result += new_set.size() - old_set.size()
+
+        return result
+
 # IC Model
 cdef class IndependentCascadeModel(DiffusionModel):
     # Functions that interface with the Python side of things
     def __cinit__(
-            self,
-            array.array starts,
-            array.array edges,
-            *,
-            double activation_prob = 0.1,
-            array.array activation_probs = None,
-            array.array _edge_probabilities = None
-        ):
+        self,
+        unsigned int[:] starts not None,
+        unsigned int[:] edges not None,
+        *,
+        double activation_prob = 0.1,
+        float[:] activation_probs = None,
+        float[:] payoffs = None,
+        float[:] _edge_probabilities = None
+    ):
 
         self.starts = starts
         self.edges = edges
-        self._edge_probabilities = _edge_probabilities
         self.activation_prob = activation_prob
         self.activation_probs = activation_probs
+        self.payoffs = payoffs
+
+        self._edge_probabilities = _edge_probabilities
 
         if self._edge_probabilities is not None:
             assert len(self.edges) == len(self._edge_probabilities)
@@ -63,13 +82,18 @@ cdef class IndependentCascadeModel(DiffusionModel):
         if self.activation_probs is not None:
             assert len(self.edges) == len(self.activation_probs)
 
+        if self.payoffs is not None:
+            assert len(self.starts) == len(self.payoffs)
+
     def set_seeds(self, seeds):
         self.original_seeds.clear()
         n = len(self.starts)
 
         for seed in seeds:
             if not (isinstance(seed, int) and 0 <= seed < n):
-                raise ValueError(f"Invalid seed node: {seed}. Must be in the range [0, {n-1}]")
+                raise ValueError(
+                    f"Invalid seed node: {seed}. Must be in the range [0, {n-1}]"
+                )
             self.original_seeds.insert(seed)
 
         self.reset_model()
@@ -90,7 +114,77 @@ cdef class IndependentCascadeModel(DiffusionModel):
     def get_num_activated_nodes(self):
         return self.seen_set.size()
 
-    cdef inline int __activation_succeeds(self, unsigned int edge_idx) except -1 nogil:
+    def compute_marginal_gain(self, seed_set, new_seed, num_trials):
+        cdef cset[unsigned int] original_seeds
+        n = len(self.starts)
+
+        for seed in seed_set:
+            if not (isinstance(seed, int) and 0 <= seed < n):
+                raise ValueError(
+                    f"Invalid seed node: {seed}. Must be in the range [0, {n-1}]"
+                )
+            elif seed == new_seed:
+                raise ValueError(
+                    f"new_seed {new_seed} should not be contained within the seed set."
+                )
+            original_seeds.insert(seed)
+
+        if new_seed is None:
+            new_seed = n
+            # Special value to compute marginal gain differently.
+        else:
[…]
