+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "cynetdiff"
-version = "0.1.13"
+version = "0.1.14"
 description = "A fast network diffusion library"
 authors = ["Eliot W. Robson <eliot.robson24@gmail.com>"]
 license = "MIT"
@@ -19,6 +19,7 @@ classifiers = [
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Topic :: Scientific/Engineering :: Mathematics",
     "Intended Audience :: Science/Research",
 ]
@@ -56,6 +57,7 @@ ruff = "^0.3.0"
 pandas-stubs = "^2.2.0.240218"
 types-tqdm = "^4.66.0.20240106"
 cython = "^3.0.9"
+cython-lint = "^0.16.2"
 
 [tool.poetry.group.docs]
 optional = true
@@ -69,10 +71,8 @@ mkdocstrings-python = "^1.7.3"
 setuptools = "^69.0.3"
 
 
-# NOTE to enable Cython generation during development, add "cython" to the
-# build dependencies before doing "poetry install".
 [build-system]
-requires = ["poetry-core", "setuptools"]
+requires = ["poetry-core", "setuptools", "cython"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.build]
@@ -81,7 +81,7 @@ generate-setup-file = false
 
 
 [tool.mypy]
-python_version = "3.10"
+python_version = "3.13"
 
 [[tool.mypy.overrides]]
 module = ["networkx.*", "Cython.*", "ndlib.*", "pooch.*", "matplotlib.*"]
