+++ .github/workflows/tests.yml
@@ -15,7 +15,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.9", "3.10", "3.11", "3.12"]
+        python-version: ["3.9", "3.10", "3.11", "3.12", "3.13"]
         poetry-version: ["1.8.1"]
 
     steps:
@@ -33,10 +33,5 @@ jobs:
       - name: Install dependencies
         run: poetry install --with dev
 
-      #- name: Install dependencies
-      #  run: |
-      #    python -m pip install -r build_requirements.txt
-      #    python -m pip install -r test_requirements.txt
-      #    python setup.py build_ext --inplace
       - name: Test with pytest
         run: poetry run pytest

[…]
