### 2024-10-29 #175 @sneakers-the-rat
Catching up - here are the things i see raised in the review:

making a checklist mostly for tidyiness sake, not trying to list out requirements - tell me which of these are taken care of/still outstanding

- [ ] docs: examples
- [ ] statement of need: we don't have a hard statement of need requirement, and i don't like novelty tests so it's not that, but it would be nice to have some statement for the pyos site that puts this package in context with other packages - "similar to x except y, etc." kind of stuff
- [ ] packaging: wheels for multiple platforms (looks ongoing)
- [x] packaging: versioning generated artifacts (looks like this is done?)
- [ ] function: use of c's rand function (looks like this is ongoing)
- [ ] function: name mangling

Thanks for the detailed review :) I learned some things about cython here too.

If it would be helpful, you might consider opening up the remaining items as issues that link back to this one just to keep track of what's left todo/use issues as a cheap form of threading.

Re: building for multiple platforms, this sounds like something we would want to have in the packaging guide, so if you figure something out, maybe we put instructions there?
