### 2024-10-29 #175 @eliotwrobson
@sneakers-the-rat thanks for writing that out. I've removed all the instances of name mangling I could find, and you are correct that removing generated artifacts is done. For the statement of need, would you like something written that can go on the pyos site, or that also to be added to the library documentation? A simple description is that this package is a faster version of [ndlib](https://ndlib.readthedocs.io/en/latest/).
