### 2024-10-29 #175 @Kai-Striega
> Re: building for multiple platforms, this sounds like something we would want to have in the packaging guide, so if you figure something out, maybe we put instructions there?

While this is noble, I think a unified approach to building will be hard, especially when extension modules are involved. It will depend on which languages are used, which platforms you build for, how complicated you want to get, and so on...

In particular, I feel that Poetry doesn't handle extension modules particularly well. There is [SPIN](https://github.com/scientific-python/spin) which much of the scientific Python ecosystem seems to be standardising on, however this would require a complete rewrite of CyNetDiff's infrastructure.

> A simple description is that this package is a faster version of [ndlib](https://ndlib.readthedocs.io/en/latest/).

If you're going to make claims that it's faster, it would be great to show it. I think another point would be the lack of benchmarks that @sneakers-the-rat missed in his list. I think this should be required if you make performance claims.
