### 2024-10-30 #175 @sneakers-the-rat
If you have published benchmarks already, then that's good enough for me for a claim like that - we can just cite them :)

Also am i reading this right, that an identical operation took 14s with ndlib and 0.067s with cynetdiff? that's.... certainly a perf improvement. well done.

Also i forgot to say earlier, i was asking for a context statement for the website but i also generally (i.e. outside of the purposes of the review) think they are good to have in documentation. If you think you've improved on something, I don't think it's disrespectful to compare yourself to it esp on something objective like 'the time it takes to perform an identical operation,' and it's helpful to everyone to have better linking between what's out there. One example i love is how the first two sections of the dask dataframes ['best practices' page](https://docs.dask.org/en/stable/dataframe-best-practices.html) are "use pandas" (and then clarifications about when you would want to use dask)
