### 2024-11-01 #175 @Kai-Striega
@eliotwrobson

> For the benchmark, some light benchmarks are included in this example notebook: https://github.com/eliotwrobson/CyNetDiff/blob/main/examples/visualizations.ipynb

I'm happy with the benchmarks as-is, for the purpose of a formal review, I think there is room for improvement.

> I'm not sure if there's a more standard way of doing this? We also include some discussion of these benchmarks in our paper: https://www.vldb.org/pvldb/vol17/p4409-umrawal.pdf

There are tools out there to do this for you. The one that seems to be favoured by NumPy/SciPy (where I do most of my FOSS stuff) is [air speed velocity](https://asv.readthedocs.io/en/latest/). There are several other tools also available, but I don't know too much about them.

> Regarding building, although changing / documenting something to work with Poetry is probably not the most productive given what others have said, I think having some example related to this in the packaging guide would be extremely useful. I had a very hard time even figuring out how to get started writing a package that used Cython, and this would be extremely useful to me on another project I want to get started on before too long.

I've tried building Cython projects with Poetry and actually ended gave up. I found it very hard/confusing, so just getting your package to build should be considered an achievement. My go to for building Cython projects is now [SPIN](https://github.com/scientific-python/spin). This uses [Meson](https://mesonbuild.com/index.html) which has an inbuilt [Cython Module](https://mesonbuild.com/Cython.html#cython).
