### 2024-11-04 #175 @eliotwrobson
> I've tried building Cython projects with Poetry and actually gave up. I found it very hard/confusing, so just getting your package to build should be considered an achievement. My go to for building Cython projects is now [SPIN](https://github.com/scientific-python/spin). This uses [Meson](https://mesonbuild.com/index.html) which has an inbuilt [Cython Module](https://mesonbuild.com/Cython.html#cython).

@sneakers-the-rat I think that even this simple recommendation alone would be a useful thing to add to the packaging guide 👍🏽

Regarding the remaining review items, I need to refactor another feature for this package to implement something for a paper I'm working on. I think the other reviewer is going to be able to take a look soon, so I'll wait for his comments before attacking the rest of these items.
