+++ src/cynetdiff/models.pyx
@@ -114,7 +114,7 @@ cdef class IndependentCascadeModel(DiffusionModel):
         return self.seen_set.size()
 
     def compute_marginal_gains(self, seed_set, new_seeds, num_trials):
-        cdef cset[unsigned int] original_seeds
+        cdef cvector[unsigned int] original_seeds
         cdef cvector[unsigned int] new_seeds_vec
         n = len(self.starts)
 
@@ -132,9 +132,9 @@ cdef class IndependentCascadeModel(DiffusionModel):
                 )
             elif seed in new_seeds_set:
                 raise ValueError(
-                    f"New seed {new_seed} should not be contained within the seed set."
+                    f"New seed {seed} should not be contained within the seed set."
                 )
-            original_seeds.insert(seed)
+            original_seeds.push_back(seed)
 
         for new_seed in new_seeds:
             if not (isinstance(new_seed, int) and 0 <= new_seed < n):
@@ -147,16 +147,14 @@ cdef class IndependentCascadeModel(DiffusionModel):
             original_seeds, new_seeds_vec, num_trials
         )
 
-        res_list = []
+        return [
+            num for num in results
+        ]
 
-        for res_number in results:
-            res_list.append(res_number)
-
-        return res_list
 
     cdef cvector[float] _compute_marginal_gains(
         self,
-        cset[unsigned int]& original_seeds,
+        cvector[unsigned int]& original_seeds,
         cvector[unsigned int]& new_seeds,
         unsigned int num_trials
     ):
@@ -181,7 +179,7 @@ cdef class IndependentCascadeModel(DiffusionModel):
                 self._advance_model(work_deque, seen_set)
 
             for i in range(1, new_seeds.size()+1):
-                new_seed = new_seeds[i]
+                new_seed = new_seeds[i-1]
 
                 # No marginal gain unless we're activating a new node
                 if seen_set.find(new_seed) == seen_set.end():
@@ -189,8 +187,8 @@ cdef class IndependentCascadeModel(DiffusionModel):
                     seen_set.insert(new_seed)
 
                     while work_deque.size() > 0:
-                        self._advance_model(work_deque, seen_set)
                         results[i] += self._compute_payoff(work_deque, self.payoffs)
+                        self._advance_model(work_deque, seen_set)
 
         for i in range(results.size()):
             results[i] /= num_trials
@@ -262,247 +260,247 @@ cdef class IndependentCascadeModel(DiffusionModel):
 
 
 # LT Model
-cdef class LinearThresholdModel(DiffusionModel):
-    # Functions that interface with the Python side of things
-    def __cinit__(
-        self,
-        unsigned int[:] starts not None,
-        unsigned int[:] edges not None
-        *,
-        float[:] influence = None,
-        float[:] payoffs = None,
-    ):
-        self.starts = starts
-        self.edges = edges
-        self.payoffs = payoffs
+# cdef class LinearThresholdModel(DiffusionModel):
+#     # Functions that interface with the Python side of things
+#     def __cinit__(
+#         self,
+#         unsigned int[:] starts not None,
+#         unsigned int[:] edges not None
+#         *,
+#         float[:] influence = None,
+#         float[:] payoffs = None,
+#     ):
+#         self.starts = starts
+#         self.edges = edges
+#         self.payoffs = payoffs
 
-        cdef unsigned int n = len(self.starts)
-        cdef unsigned int m = len(self.edges)
-        cdef cvector[unsigned int] in_degrees
+#         cdef unsigned int n = len(self.starts)
+#         cdef unsigned int m = len(self.edges)
+#         cdef cvector[unsigned int] in_degrees
 
-        # Setting the influence sent across each edge
-        if influence is not None:
-            # If provided, copy from user code
-            assert m == len(influence)
[…]
