### 2024-11-05 6110a38
Refactored linear threshold tests (mostly)
