+++ src/cynetdiff/models.pyx
@@ -37,6 +37,8 @@ cdef class DiffusionModel:
     cpdef void advance_until_completion(self):
         raise NotImplementedError
 
+    @cython.boundscheck(False)
+    @cython.wraparound(False)
     cdef float _compute_payoff(
         self,
         cdeque[unsigned int]& activated_nodes,
@@ -150,7 +152,8 @@ cdef class IndependentCascadeModel(DiffusionModel):
             num for num in results
         ]
 
-
+    @cython.boundscheck(False)
+    @cython.wraparound(False)
     cdef cvector[float] _compute_marginal_gains(
         self,
         cvector[unsigned int]& original_seeds,
@@ -160,13 +163,11 @@ cdef class IndependentCascadeModel(DiffusionModel):
         cdef cdeque[unsigned int] work_deque
         cdef cset[unsigned int] seen_set
 
-        cdef float result = 0.0
-        cdef unsigned int n = len(self.starts)
-
         cdef cvector[float] results
         results = cvector[float](new_seeds.size()+1, 0.0)
 
         cdef unsigned int new_seed
+        cdef unsigned int i
 
         for _ in range(num_trials):
             work_deque.assign(original_seeds.begin(), original_seeds.end())
@@ -190,7 +191,7 @@ cdef class IndependentCascadeModel(DiffusionModel):
                         self._advance_model(work_deque, seen_set)
 
         for i in range(results.size()):
-            results[i] /= num_trials
+            results[i] /= <float>num_trials
 
         return results
 
@@ -259,247 +260,261 @@ cdef class IndependentCascadeModel(DiffusionModel):
 
 
 # LT Model
-# cdef class LinearThresholdModel(DiffusionModel):
-#     # Functions that interface with the Python side of things
-#     def __cinit__(
-#         self,
-#         unsigned int[:] starts not None,
-#         unsigned int[:] edges not None
-#         *,
-#         float[:] influence = None,
-#         float[:] payoffs = None,
-#     ):
-#         self.starts = starts
-#         self.edges = edges
-#         self.payoffs = payoffs
+cdef class LinearThresholdModel(DiffusionModel):
+    # Functions that interface with the Python side of things
+    def __cinit__(
+        self,
+        unsigned int[:] starts not None,
+        unsigned int[:] edges not None
+        *,
+        float[:] influence = None,
+        float[:] payoffs = None,
+    ):
+        self.starts = starts
+        self.edges = edges
+        self.payoffs = payoffs
 
-#         cdef unsigned int n = len(self.starts)
-#         cdef unsigned int m = len(self.edges)
-#         cdef cvector[unsigned int] in_degrees
+        cdef unsigned int n = len(self.starts)
+        cdef unsigned int m = len(self.edges)
+        cdef cvector[unsigned int] in_degrees
 
-#         # Setting the influence sent across each edge
-#         if influence is not None:
-#             # If provided, copy from user code
-#             assert m == len(influence)
-#             self.influence = influence
-#         else:
-#             # Otherwise, default to 1/in_degree
-#             in_degrees.resize(n)
-#             fill(in_degrees.begin(), in_degrees.end(), 0)
+        # Setting the influence sent across each edge
+        if influence is not None:
+            # If provided, copy from user code
+            assert m == len(influence)
+            self.influence = influence
+        else:
+            # Otherwise, default to 1/in_degree
+            in_degrees.resize(n)
+            fill(in_degrees.begin(), in_degrees.end(), 0)
 
-#             for out_node in self.edges:
-#                 in_degrees[out_node] += 1
+            for out_node in self.edges:
+                in_degrees[out_node] += 1
 
-#             influence_arr = array.array("f")
-#             influence_arr.extend(
-#                 1.0 / in_degrees[out_node]
-#                 for out_node in self.edges
-#             )
[…]
