### 2024-11-05 96dee52
Added newer citation
