### 2024-11-05 57ffa22
Marginal gain list update (#24)
