+++ src/cynetdiff/models.pyx
@@ -37,20 +37,20 @@ cdef class DiffusionModel:
     cpdef void advance_until_completion(self):
         raise NotImplementedError
 
+    @cython.boundscheck(False)
+    @cython.wraparound(False)
     cdef float _compute_payoff(
         self,
-        cset[unsigned int]& new_set,
-        cset[unsigned int]& old_set,
+        cdeque[unsigned int]& activated_nodes,
         float[:] payoffs,
     ):
         cdef float result = 0.0
 
         if payoffs is not None:
-            for node in new_set:
-                if old_set.find(node) == old_set.end():
-                    result += payoffs[node]
+            for node in activated_nodes:
+                result += payoffs[node]
         else:
-            result += new_set.size() - old_set.size()
+            result += <float>activated_nodes.size()
 
         return result
 
@@ -114,46 +114,61 @@ cdef class IndependentCascadeModel(DiffusionModel):
     def get_num_activated_nodes(self):
         return self.seen_set.size()
 
-    def compute_marginal_gain(self, seed_set, new_seed, num_trials):
-        cdef cset[unsigned int] original_seeds
+    def compute_marginal_gains(self, seed_set, new_seeds, num_trials):
+        cdef cvector[unsigned int] original_seeds
+        cdef cvector[unsigned int] new_seeds_vec
         n = len(self.starts)
+
+        new_seeds_set = set(new_seeds)
+
+        if len(new_seeds) != len(new_seeds_set):
+            raise ValueError(
+                "New seeds set must have all unique elements."
+            )
 
         for seed in seed_set:
             if not (isinstance(seed, int) and 0 <= seed < n):
                 raise ValueError(
                     f"Invalid seed node: {seed}. Must be in the range [0, {n-1}]"
                 )
-            elif seed == new_seed:
+            elif seed in new_seeds_set:
                 raise ValueError(
-                    f"new_seed {new_seed} should not be contained within the seed set."
+                    f"New seed {seed} should not be contained within the seed set."
                 )
-            original_seeds.insert(seed)
+            original_seeds.push_back(seed)
 
-        if new_seed is None:
-            new_seed = n
-            # Special value to compute marginal gain differently.
-        else:
+        for new_seed in new_seeds:
             if not (isinstance(new_seed, int) and 0 <= new_seed < n):
                 raise ValueError(
-                    f"Invalid new_seed: {new_seed}. Must be in the range [0, {n-1}]"
+                    f"Invalid new seed {new_seed}. "
+                    f"Must be integer in the range [0, {n-1}]"
                 )
+            new_seeds_vec.push_back(new_seed)
 
-        return self._compute_marginal_gain(
-            original_seeds, new_seed, num_trials
+        cdef cvector[float] results = self._compute_marginal_gains(
+            original_seeds, new_seeds_vec, num_trials
         )
 
-    cdef float _compute_marginal_gain(
+        return [
+            num for num in results
+        ]
+
+    @cython.boundscheck(False)
+    @cython.wraparound(False)
+    cdef cvector[float] _compute_marginal_gains(
         self,
-        cset[unsigned int]& original_seeds,
-        unsigned int new_seed,
+        cvector[unsigned int]& original_seeds,
+        cvector[unsigned int]& new_seeds,
         unsigned int num_trials
     ):
         cdef cdeque[unsigned int] work_deque
         cdef cset[unsigned int] seen_set
-        cdef cset[unsigned int] new_seen_set
 
-        cdef float result = 0.0
-        cdef unsigned int n = len(self.starts)
+        cdef cvector[float] results
+        results = cvector[float](new_seeds.size()+1, 0.0)
+
+        cdef unsigned int new_seed
+        cdef unsigned int i
 
         for _ in range(num_trials):
[…]
