### 2024-11-05 eliotwrobson/CyNetDiff#25 @eliotwrobson
Add examples for user facing functions
From the review [here](https://github.com/pyOpenSci/software-submission/issues/175#issuecomment-2439393615), add (more) examples for user facing functions in the documentation and `models.pyi` file.
