### 2024-11-05 eliotwrobson/CyNetDiff#26 @eliotwrobson
Add benchmarks
Add benchmarks using ASV, from discussion [here](https://github.com/pyOpenSci/software-submission/issues/175#issuecomment-2452697024).
