### 2024-11-05 eliotwrobson/CyNetDiff#27 @eliotwrobson
Provide wheels for other platforms
See discussion [here](https://github.com/pyOpenSci/software-submission/issues/175#issuecomment-2439393615).
