### 2024-11-16 #175 @jeremyzhangsq
@eliotwrobson check this out:D

- [x]  As the reviewer I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

- [x]  Package has a [[README.md](http://readme.md/)](http://readme.md/) file in the root directory.

    - [x]  A [[repostatus.org](https://www.repostatus.org/)](https://www.repostatus.org/) badge,
        - [x]  Yes for PyPI

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [[see this example](https://github.com/ropensci/drake)](https://github.com/ropensci/drake). Such a table should be more wide than high. (Note that the a badge for pyOpenSci peer-review will be provided upon acceptance.)*

- [x]  Descriptive links to all vignettes. If the package is small, there may only be a need for one vignette which could be placed in the [[README.md](http://readme.md/)](http://readme.md/) file.
    - [x]  The authors have compared this package with NDlib, a wide-accepted library for network diffusion. This package is more efficient (See the PVLDB paper).

### Usability

    - [x]  Although the authors provided the benchmark codes, I suggest the authors give some key results in the README.
- [x]  **Packaging guidelines**: The package conforms to the pyOpenSci [[packaging guidelines](https://www.pyopensci.org/python-package-guide)](https://www.pyopensci.org/python-package-guide).
    - [x]  Package supports modern versions of Python and not [[End of life versions](https://endoflife.date/python)](https://endoflife.date/python).

### For packages also submitting to JOSS

- [ ]  The package has an **obvious research application** according to JOSS's definition in their [[submission requirements](http://joss.theoj.org/about#submission_requirements)](http://joss.theoj.org/about#submission_requirements).

The package contains a `paper.md` matching [[JOSS's requirements](http://joss.theoj.org/about#paper_structure)](http://joss.theoj.org/about#paper_structure) with:

### Final approval (post-review)

---

I have to admit that this project really surprised me. Previously, the IM community mainly used C++ for implementation for the sake of efficiency. However, it may cause inconvenience for real-world development and deployment (from my own experience). Thus, I believe this project will help our community a lot. Thanks to other reviewers’ comments, I can quickly catch up with the progress. Overall, this project meets my standards except two minors to improve.

- I suggest showing some key evaluation results in the README, such as how fast CyNetDiff can be compared to NDlib.
- To adapt other variants of IC and LT models in the future, I suggest to provide args about initial node thresholds and edge weights in

```
def linear_threshold(
    G: t.Union[nx.Graph, nx.DiGraph], seeds: t.Iterable[int], steps: int = 0
) -> t.List[t.List[int]]:
```

An example of the LT variant that requires customized thresholds and weights is this:

- Nicola Barbieri and Francesco Bonchi. 2014. Influence maximization with viral
product design. In ICDM.
