+++ pyproject.toml
@@ -72,7 +72,7 @@ setuptools = "^69.0.3"
 
 
 [build-system]
-requires = ["poetry-core", "setuptools", "cython"]
+requires = ["poetry-core", "setuptools", "cython", "numpy"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.build]
