### 2024-12-14 30bf7ec
Modify RNG
