### 2024-12-14 850948b
Simplified type hints
