### 2024-12-14 #175 @eliotwrobson
Finally had a chance to go over things again, and I managed to get a PR for the most significant item so far (SPEC7 compliance): https://github.com/eliotwrobson/CyNetDiff/pull/28

@Kai-Striega if you have time, could you glance at this PR briefly? No need to spend too much time looking through the internals, I just want to double check that we're doing everything as expected in the spec.

As for some of the other comments, I think that we have a sufficient amount of customization in the model creation for now (edge weights are accepted as arguments by both models, and adding custom thresholds for the linear threshold model won't give a performance boost over pure-python unless we adopt a custom model of randomly generating thresholds in each simulation). So the only remaining items are the following

- [ ] More examples in the documentation.
- [ ] ASV / something showing more current benchmarks.
- [ ] Releasing wheels on multiple platforms (I probably won't do this because I can't figure out how).
