### 2024-12-14 #175 @Kai-Striega
@eliotwrobson I'll take a look, hopefully today.

> Releasing wheels on multiple platforms (I probably won't do this because I can't figure out how).

I think having wheels for multiple platforms is important. If you don't supply them, the user has to figure out how to compile your project on their own. i.e. you're moving the burden of having to figure out how to compile it on their platform from yourself to each individual user.

I'm happy to help with this a bit if you need it, basically for NumPy/SciPy we build it as a CI job. You can see how NumPy does it by looking in the [.github/workflows/](https://github.com/numpy/numpy/tree/main/.github/workflows) directory.
