### 2024-12-14 #175 @eliotwrobson
> I'm happy to help with this a bit if you need it, basically for NumPy/SciPy we build it as a CI job. You can see how NumPy does it by looking in the [.github/workflows/](https://github.com/numpy/numpy/tree/main/.github/workflows) directory.

@Kai-Striega Thank you! This would be greatly appreciated 🙏🏽 🙏🏽 I'd be open to switching the build system as well if that is necessary, I'm just currently not sure how to set up automated publishing with respect to this.
