### 2024-12-14 eliotwrobson/CyNetDiff#27 @Kai-Striega
@eliotwrobson I'm going to bring the discussion about this here, rather than spamming the other reviewers.

I'll dig into this for a bit, but I won't have the time to look into it today (I'll look at the SPEC-7 stuff), or for the next few days. But I will get to it once I have time.

> Thank you! This would be greatly appreciated 🙏🏽 🙏🏽

You're welcome.

> I'd be open to switching the build system as well if that is necessary,

I don't think it will be _necessary_, but I think it will be better. From my experience so far, Poetry doesn't work that well with Cython. (unfortunately, it's a great tool otherwise). I normally recommend [spin](https://github.com/scientific-python/spin), but that means you'd have to learn (some) meson.

> I'm just currently not sure how to set up automated publishing with respect to this.

spin has a ``build`` command that builds your library for you. However, in the documentation it only lists ``sdist``. I'll need to see how that works. It might work out of the box, otherwise, you can add a [custom command](https://github.com/scientific-python/spin?tab=readme-ov-file#-custom-commands) to build it. I'll reach out to some people who might be able to figure it out.
