### 2024-12-14 eliotwrobson/CyNetDiff#27 @eliotwrobson
> I don't think it will be _necessary_, but I think it will be better. From my experience so far, Poetry doesn't work that well with Cython. (unfortunately, it's a great tool otherwise). I normally recommend [spin](https://github.com/scientific-python/spin), but that means you'd have to learn (some) meson.

No worries about learning meson, I'm totally open to switching tools. I'm going to start another project before too long that will be using Cython as well, so being able to learn / copy over infrastructure from this project would make that much easier later.

I'm also occupied with a couple of other things at the moment (including the somewhat more pressing feature releases here), so no rush 👍🏽
