### 2024-12-14 eliotwrobson/CyNetDiff#27 @Kai-Striega
I think what you want is the [cibuildwheel](https://github.com/pypa/cibuildwheel) project.
