### 2024-12-15 3941550
Update utils.py
