+++ tests/test_lt_model.py
@@ -6,6 +6,7 @@ import typing as t
 
 import networkx as nx
 import pytest
+
 from cynetdiff.utils import networkx_to_lt_model
 
 #    Code below adapted from code by
@@ -198,6 +199,33 @@ def get_thresholds(graph: nx.DiGraph) -> array.array:
 
 
 # Start of actual test code
+
+
+@pytest.mark.parametrize("directed", [True, False])
+@pytest.mark.parametrize("seed", [12345, 505050, 2024])
+def test_model_rng_seed(directed: bool, seed: int) -> None:
+    n = 10000
+    k = 10
+    p = 0.01
+    # Just trying the main functions with no set thresholds
+    graph = generate_random_graph_from_seed(n, p, directed, False, False, seed=seed)
+    model, _ = networkx_to_lt_model(graph)
+    model.set_rng(seed)
+
+    random.seed(seed)
+    seeds = set(random.sample(list(graph.nodes), k))
+    model.set_seeds(seeds)
+    model.advance_until_completion()
+
+    activated_nodes = set(model.get_activated_nodes())
+
+    # Reset model and run again
+    model.reset_model()
+    model.set_seeds(seeds)
+    model.set_rng(seed)
+    model.advance_until_completion()
+
+    assert activated_nodes == set(model.get_activated_nodes())
 
 
 @pytest.mark.parametrize("directed", [True, False])
