### 2024-12-29 eebc48b
Improved RNG test case
