### 2024-12-29 84005e6
Make library SPEC-7 compliant (#28)
