+++ tests/test_ic_model.py
@@ -5,6 +5,7 @@ import typing as t
 
 import networkx as nx
 import pytest
+
 from cynetdiff.utils import (
     networkx_to_ic_model,
     set_activation_random_sample,
@@ -167,13 +168,15 @@ def generate_random_graph_from_seed(
     directed: bool,
     include_payoff: bool,
     *,
+    include_success_prob: bool = True,
     seed: int = 12345,
 ) -> t.Union[nx.Graph, nx.DiGraph]:
     graph = nx.fast_gnp_random_graph(n, p, directed=directed, seed=seed)
 
     random.seed(seed)
-    for _, _, data in graph.edges(data=True):
-        data["success_prob"] = random.random()
+    if include_success_prob:
+        for _, _, data in graph.edges(data=True):
+            data["success_prob"] = random.random()
 
     if include_payoff:
         for _, data in graph.nodes(data=True):
@@ -183,6 +186,47 @@ def generate_random_graph_from_seed(
 
 
 # The start of the actual test cases
+
+
+@pytest.mark.parametrize("directed", [True, False])
+@pytest.mark.parametrize("seed", [12345, 505050, 2024])
+def test_model_rng_seed(directed: bool, seed: int) -> None:
+    n = 10000
+    k = 10
+    p = 0.005
+    num_runs = 10
+    # Just trying the main functions with no set thresholds
+    graph = generate_random_graph_from_seed(
+        n, p, directed, False, seed=seed, include_success_prob=False
+    )
+    model, _ = networkx_to_ic_model(graph, rng=seed)
+
+    random.seed(seed)
+    seeds = set(random.sample(list(graph.nodes), k))
+
+    model.set_seeds(seeds)
+    activated_nodes_sets = []
+
+    # Get activated nodes sets for 10 runs
+    for _ in range(num_runs):
+        model.reset_model()
+        model.advance_until_completion()
+
+        activated_nodes = set(model.get_activated_nodes())
+        activated_nodes_sets.append(activated_nodes)
+
+    # Assert that sets are different
+    model.reset_model()
+    model.advance_until_completion()
+    assert activated_nodes != set(model.get_activated_nodes())
+
+    # Reseed model and run again
+    model.set_rng(seed)
+
+    for activated_nodes in activated_nodes_sets:
+        model.reset_model()
+        model.advance_until_completion()
+        assert activated_nodes == set(model.get_activated_nodes())
 
 
 @pytest.mark.parametrize("directed", [True, False])
