+++ pyproject.toml
@@ -53,6 +53,7 @@ pytest = "^8.0.2"
 mypy = "^1.7.1"
 types-setuptools = "^69.0.0.0"
 networkx = "^3.2.1"
+numpy = "^2.0.0"
 ruff = "^0.3.0"
 pandas-stubs = "^2.2.0.240218"
 types-tqdm = "^4.66.0.20240106"
@@ -72,7 +73,7 @@ setuptools = "^69.0.3"
 
 
 [build-system]
-requires = ["poetry-core", "setuptools", "cython"]
+requires = ["poetry-core", "setuptools", "cython", "numpy"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.build]

[…]
