### 2024-12-29 eliotwrobson/CyNetDiff#26 @eliotwrobson
Looking at various options, I think I will opt to add benchmarks that are run as a workflow using pytest benchmark. I don't think this project has quite the size for something like ASV to be super necessary (there are only a couple of operations we care about, and these are all being done with the same code). We can always go back and add this later, but frankly I don't expect the code to change that much once the PyOS review is done.
