+++ .github/workflows/benchmark.yml
@@ -1,5 +1,5 @@
 # This workflow will install Python dependencies, run benchmarks
-name: tests
+name: benchmarks
 
 on:
   push:
