+++ .github/workflows/tests.yml
@@ -5,7 +5,7 @@ name: tests
 
 on:
   push:
-    branches: ["*"]
+    branches: ["main"]
   pull_request:
     branches: ["*"]
 
