+++ .github/workflows/benchmark.yml
@@ -13,7 +13,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.13"]
+        python-version: ["3.12"]
         poetry-version: ["1.8.5"]
 
     steps:
