### 2024-12-30 cc2e23e
Add benchmarks using pytest-benchmark (#29)
