+++ .github/workflows/lint.yml
@@ -5,7 +5,7 @@ name: lint
 
 on:
   push:
-    branches: ["*"]
+    branches: ["main"]
   pull_request:
     branches: ["*"]
 
