+++ .github/workflows/benchmark.yml
@@ -1,7 +1,5 @@
-# This workflow will install Python dependencies, run tests and lint with a variety of Python versions
-# For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
-
-name: tests
+# This workflow will install Python dependencies, run benchmarks
+name: benchmarks
 
 on:
   push:
@@ -15,8 +13,8 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.9", "3.10", "3.11", "3.12", "3.13"]
-        poetry-version: ["1.8.1"]
+        python-version: ["3.12"]
+        poetry-version: ["1.8.5"]
 
     steps:
       - uses: actions/checkout@v4
@@ -31,7 +29,7 @@ jobs:
           poetry-version: ${{ matrix.poetry-version }}
 
       - name: Install dependencies
-        run: poetry install --with dev
+        run: poetry install --with dev,benchmark
 
-      - name: Test with pytest
-        run: poetry run pytest
+      - name: Benchmark with pytest-benchmark
+        run: poetry run pytest --benchmark-only benchmarks/test_benchmark.py
