+++ pyproject.toml
@@ -28,13 +28,11 @@ classifiers = [
 homepage = "https://github.com/eliotwrobson/CyNetDiff"
 tracker = "https://github.com/eliotwrobson/CyNetDiff/issues"
 documentation = "https://eliotwrobson.github.io/CyNetDiff/"
-# TODO add these later
-#gitter = "https://gitter/area"
 
 [tool.poetry.dependencies]
 python = ">=3.9"
 
-# TODO change the group name to "examples"
+
 [tool.poetry.group.benchmark]
 optional = true
 
