+++ examples/demo_backend.py
@@ -13,14 +13,15 @@ from functools import wraps
 
 import matplotlib.pyplot as plt
 import ndlib.models.epidemics as ep
-import ndlib.models.ModelConfig as mc
+import ndlib.models.ModelConfig as Mc
 import networkx as nx
 import numpy as np
 import pooch
 from coloraide import Color
+from tqdm.notebook import tqdm, trange
+
 from cynetdiff.models import DiffusionModel
 from cynetdiff.utils import networkx_to_ic_model
-from tqdm.notebook import tqdm, trange
 
 DiffusionGraphT = t.Union[nx.Graph, nx.DiGraph]
 SeedSetT = set[int]
@@ -47,13 +48,13 @@ def timing(f: t.Callable) -> t.Callable:
 # Pure Python diffusion code
 
 
-def independent_cascade(G: DiffusionGraphT, seeds: SeedSetT) -> list[list[int]]:
+def independent_cascade(graph: DiffusionGraphT, seeds: SeedSetT) -> list[list[int]]:
     """
     A basic pure-python implementation of independent cascade. Optimized so we can get
     a baseline reading on the best possible speed for this algorithm in pure-Python.
     """
-    if not G.is_directed():
-        G = G.to_directed()
+    if not graph.is_directed():
+        graph = graph.to_directed()
 
     visited = set(seeds)
 
@@ -67,7 +68,7 @@ def independent_cascade(G: DiffusionGraphT, seeds: SeedSetT) -> list[list[int]]:
         current_layer = []
 
         for next_node in res[-1]:
-            for child, data in G[next_node].items():
+            for child, data in graph[next_node].items():
                 if child not in visited:
                     # Lazy getter to deal with not having this set but still being
                     # efficient
@@ -109,7 +110,7 @@ def diffuse_python(
 def diffuse_ndlib(graph: DiffusionGraphT, seeds: SeedSetT, num_samples: int) -> float:
     model = ep.IndependentCascadesModel(graph)
 
-    config = mc.Configuration()
+    config = Mc.Configuration()
 
     # Assume that thresholds were already set.
     for u, v, data in graph.edges(data=True):
@@ -141,9 +142,7 @@ def diffuse_ndlib(graph: DiffusionGraphT, seeds: SeedSetT, num_samples: int) -> 
 
 
 @timing
-def diffuse_CyNetDiff(
-    graph: DiffusionGraphT, seeds: SeedSetT, num_samples: int
-) -> float:
+def diffuse_cynetdiff(graph: DiffusionGraphT, seeds: SeedSetT, num_samples: int) -> float:
     model, _ = networkx_to_ic_model(graph)
     model.set_seeds(seeds)
 
@@ -165,16 +164,19 @@ def simple_benchmark(
     *,
     num_seeds: int = 10,
     num_trials: int = 1_000,
-    backends_to_run: t.List[MethodsT] = [
-        "cynetdiff",
-        "ndlib",
-        "python",
-    ],
+    backends_to_run: t.Optional[t.List[MethodsT]] = None,
 ) -> None:
     """
     A simple benchmark function for three different diffusion
     implementations.
     """
+
+    if backends_to_run is None:
+        backends_to_run = [
+            "cynetdiff",
+            "ndlib",
+            "python",
+        ]
 
     # Number of seed nodes
     print(f"Number of randomly chosen seed nodes: {num_seeds:,d}")
@@ -193,7 +195,7 @@ def simple_benchmark(
 
     if "cynetdiff" in backends_to_run:
         print("Starting diffusion with CyNetDiff.")
-        avg_cynetdiff = diffuse_CyNetDiff(graph, seeds, num_trials)
+        avg_cynetdiff = diffuse_cynetdiff(graph, seeds, num_trials)
 
[…]
