### 2024-12-30 952b09b
More restrictive formatting (#30)
