+++ tests/test_ic_model.py
@@ -23,7 +23,7 @@ Graph = t.Union[nx.Graph, nx.DiGraph]
 
 
 def independent_cascade(
-    G: Graph,
+    graph: Graph,
     seeds: t.Iterable[int],
     *,
     steps: int = 0,
@@ -31,60 +31,51 @@ def independent_cascade(
     activation_prob: t.Optional[float] = None,
 ) -> t.List[t.List[int]]:
     """Return the active nodes of each diffusion step by the independent cascade
-  model
+    model
 
-  Parameters
-  -----------
-  G : graph
-    A NetworkX graph
-  seeds : list of nodes
-    The seed nodes for diffusion
-  steps: integer
-    The number of steps to diffuse.  If steps <= 0, the diffusion runs until
-    no more nodes can be activated.  If steps > 0, the diffusion runs for at
-    most "steps" rounds
+    Parameters
+    -----------
+    G : graph
+      A NetworkX graph
+    seeds : list of nodes
+      The seed nodes for diffusion
+    steps: integer
+      The number of steps to diffuse.  If steps <= 0, the diffusion runs until
+      no more nodes can be activated.  If steps > 0, the diffusion runs for at
+      most "steps" rounds
 
-  Returns
-  -------
-  layer_i_nodes : list of list of activated nodes
-    layer_i_nodes[0]: the seeds
-    layer_i_nodes[k]: the nodes activated at the kth diffusion step
+    Returns
+    -------
+    layer_i_nodes : list of list of activated nodes
+      layer_i_nodes[0]: the seeds
+      layer_i_nodes[k]: the nodes activated at the kth diffusion step
 
-  Notes
-  -----
-  When node v in G becomes active, it has a *single* chance of activating
-  each currently inactive neighbor w with probability p_{vw}
+    Notes
+    -----
+    When node v in G becomes active, it has a *single* chance of activating
+    each currently inactive neighbor w with probability p_{vw}
 
-  Examples
-  --------
-  >>> DG = nx.DiGraph()
-  >>> DG.add_edges_from([(1,2), (1,3), (1,5), (2,1), (3,2), (4,2), (4,3), \
-  >>>   (4,6), (5,3), (5,4), (5,6), (6,4), (6,5)], activation_prob=0.2)
-  >>> layers = networkx_addon.information_propagation.independent_cascade(DG, [6])
-
-  References
-  ----------
-  [1] David Kempe, Jon Kleinberg, and Eva Tardos.
-      Influential nodes in a diffusion model for social networks.
-      In Automata, Languages and Programming, 2005.
-  """
-    if isinstance(G, (nx.MultiGraph, nx.MultiDiGraph)):
-        raise Exception(
-            "independent_cascade() is not defined for graphs with multiedges."
-        )
+    References
+    ----------
+    [1] David Kempe, Jon Kleinberg, and Eva Tardos.
+        Influential nodes in a diffusion model for social networks.
+        In Automata, Languages and Programming, 2005.
+    """
+    if isinstance(graph, (nx.MultiGraph, nx.MultiDiGraph)):
+        raise Exception("independent_cascade() is not defined for graphs with multiedges.")
 
     rand_gen = random.Random(random_seed)
 
     # change to directed graph
-    if not G.is_directed():
-        DG = G.to_directed()
+    if not graph.is_directed():
+        diffusion_graph = graph.to_directed()
     else:
-        DG = copy.deepcopy(G)
+        diffusion_graph = copy.deepcopy(graph)
 
     act_prob_default = 0.1 if activation_prob is None else activation_prob
 
     # init activation probabilities
-    for u, v, data in DG.edges(data=True):
+    for _, _, data in diffusion_graph.edges(data=True):
         if "activation_prob" not in data:
[…]
