### 2024-12-30 136bec2
New payoff function
