### 2024-12-30 1c1c28b
Adding randomized activation
