### 2024-12-30 f1c8e33
feat: Randomized activation function (#31)
