+++ tests/test_ic_model.py
@@ -177,6 +177,56 @@ def generate_random_graph_from_seed(
 # The start of the actual test cases
 
 
+@pytest.mark.parametrize("seed", [12345, 505050, 2024])
+def test_randomized_activation(seed: int) -> None:
+    n = 100
+    k = 10
+    p = 0.01
+
+    # Create random graph and initialize model
+    graph = generate_random_graph_from_seed(n, p, True, False, seed=seed)
+    model, _ = networkx_to_ic_model(graph, rng=seed)
+
+    random.seed(seed)
+    seeds = random.sample(list(graph.nodes), k)
+    seed_probs = random.sample([1.0 for _ in range(k // 2)] + [0.0 for _ in range(k // 2)], k)
+
+    model.set_seeds(seeds, seed_probs)
+
+    # Check that the seeds are set correctly
+    assert model.get_num_activated_nodes() == k // 2
+
+    activated_nodes = set(model.get_activated_nodes())
+
+    for seed_node, prob in zip(seeds, seed_probs):
+        assert (seed_node in activated_nodes) == (prob == 1.0)
+
+
+@pytest.mark.parametrize("seed", [12345, 505050, 2024])
+def test_model_payoffs(seed: int) -> None:
+    n = 10_000
+    k = 10
+    p = 0.01
+
+    # Just trying the main functions with no set thresholds
+    graph = generate_random_graph_from_seed(n, p, True, True, seed=seed)
+    model, _ = networkx_to_ic_model(graph, rng=seed)
+    random.seed(seed)
+    seeds = set(random.sample(list(graph.nodes), k))
+
+    # Run the model
+    model.set_seeds(seeds)
+    model.advance_until_completion()
+    payoff_score = model.compute_payoffs()
+
+    # Compute score manually and compare
+    manual_score = 0.0
+    for node in model.get_activated_nodes():
+        manual_score += graph.nodes[node]["payoff"]
+
+    assert math.isclose(payoff_score, manual_score, abs_tol=0.3)
+
+
 @pytest.mark.parametrize("directed", [True, False])
 @pytest.mark.parametrize("seed", [12345, 505050, 2024])
 def test_model_rng_seed(directed: bool, seed: int) -> None:
