### 2024-12-30 2951ef3
Moved demo files and updated example in docs
