+++ docs/examples/celf.md
@@ -13,65 +13,9 @@ $\sigma(S)$ on the graph is maximized.
 an optimized version of the greedy algorithm. At a high level, this algorithm works by greedily adding nodes to
 the seed set which has the highest marginal gain. This is repeated until the budget is exhausted.
 
-## Computing Marginal Gains
-
-To implement this, we first need a specialized function to compute the marginal gain of adding a node to a
-given seed set. In this example, we focus on the independent cascade model.
-`CyNetDiff` allows us to accomplish this task very quickly:
-
 ```python
 import typing as t
 from cynetdiff.models import DiffusionModel
-
-def compute_marginal_gain(
-    model: DiffusionModel,
-    new_node: int,
-    seeds_list: t.List[int],
-    num_trials: int = 1_000,
-) -> float:
-    """
-    Compute the marginal gain in the spread of influence by adding a new node to the set of seed nodes,
-    by summing the differences of spreads for each trial and then taking the average.
-
-    Parameters:
-    - model: The model used for simulating the spread of influence.
-    - new_node: The new node to consider adding to the set of seed nodes.
-    - seeds: The current set of seed nodes.
-    - num_trials: The number of trials to average the spread of influence over.
-
-    Returns:
-    - The average marginal gain in the spread of influence by adding the new node.
-    """
-    seeds = set(seeds_list)
-    original_spread = 0
-    new_spread = 0
-    # If no seeds at the beginning, original spread is always just zero.
-    # Prevents wasted work in cases where the seed set is empty.
-    if len(seeds) > 0:
-        model.set_seeds(seeds)
-
-        for _ in range(num_trials):
-            model.reset_model()
-            model.advance_until_completion()
-            original_spread += model.get_num_activated_nodes()
-
-    new_seeds = seeds.union({new_node})
-    model.set_seeds(new_seeds)
-
-    for _ in range(num_trials):
-        model.reset_model()
-        model.advance_until_completion()
-        new_spread += model.get_num_activated_nodes()
-
-    # Avoid floating point division until the very end.
-    return (new_spread - original_spread) / num_trials
-```
-
-## Main Algorithm
-
-With this function, implementing the rest of the algorithm is straightforward.
-
-```python
 from tqdm import tqdm, trange
 import heapq
 
@@ -94,12 +38,7 @@ def celf(
     for node in trange(n):
         marg_gain.append(
             (
-                -compute_marginal_gain(
-                    model,
-                    node,
-                    set(),
-                    num_trials,
-                ),
+                -model.compute_marginal_gains([node], [], num_trials),
                 node,
             )
         )
@@ -108,7 +47,7 @@ def celf(
     heapq.heapify(marg_gain)
 
     max_mg, selected_node = heapq.heappop(marg_gain)
-    S = [selected_node]
+    seeds = [selected_node]
     spread = -max_mg
     spreads = [spread]
 
@@ -117,24 +56,19 @@ def celf(
     for _ in trange(k - 1):
         while True:
             current_mg, current_node = heapq.heappop(marg_gain)
-            new_mg_neg = -compute_marginal_gain(
-                model,
-                current_node,
-                S,
-                num_trials,
-            )
+            new_mg_neg = -model.compute_marginal_gains(seeds, [current_node], num_trials)[1]
 
-            if new_mg_neg > current_mg:
+            if new_mg_neg <= current_mg:
                 break
             else:
-                heapq.heappush(marg_gain, (current_mg, current_node))
+                heapq.heappush(marg_gain, (new_mg_neg, current_node))
 
         spread += -new_mg_neg
-        S.append(current_node)
+        seeds.append(current_node)
         spreads.append(spread)
 
     # Return the maximizing set S and the increasing spread values.
-    return S, spreads
+    return seeds, spreads
 ```
 
 ## Running the Algorithm
