### 2024-12-30 a60438b
Add examples of user facing functions (#32)
