### 2024-12-30 1d21361
Add CSR creation function
