### 2024-12-30 #175 @eliotwrobson
Update on the status of things here. I've made a pass and resolved all but one of the outstanding requested changes from this review (the remaining is the distribution of wheels for other platforms, which I'll be handling next). The latest released version (0.1.16) has the following changes:

- Examples for all of the user facing functions in the API, in addition to the extended examples present already: https://eliotwrobson.github.io/CyNetDiff/api/
- Benchmarks run through `pytest-benchmark` on GitHub actions for the main branch. The comparison here is done with NDlib and there is a link to the latest benchmarks run from the main branch in the README. Not the most sophisticated benchmarks in the world but appropriate given the size of the project (ASV was ultimately too much overhead when there are only really a couple of tasks we care about benchmarking here).
- Support for randomized activation and other utility functions to support additional use cases.

The wheel distribution will take a little more time than everything else, but I'll get to that as soon as I can. However, I think everything else is in a good spot 👍🏽
