+++ pyproject.toml
@@ -1,17 +1,20 @@
-[tool.poetry]
+[project]
 name = "cynetdiff"
-version = "0.1.16"
+version = "0.1.16" # TODO switch to dynamic versioning
 description = "A fast network diffusion library"
-authors = ["Eliot W. Robson <eliot.robson24@gmail.com>"]
+authors = [{ name = "Eliot W. Robson", email = "eliot.robson24@gmail.com>" }]
 license = "MIT"
+requires-python = ">=3.9"
 readme = "README.md"
-packages = [{ include = "cynetdiff", from = "src" }]
 maintainers = [
-    "Eliot W. Robson <eliot.robson24@gmail.com>",
-    "Abhishek K. Umrawal <aumrawal@illinois.edu>",
+    { name = "Eliot W. Robson", email = "eliot.robson24@gmail.com" },
+    { name = "Abhishek K. Umrawal", email = "aumrawal@illinois.edu" },
 ]
+#dependencies = ["python>=3.9"]
 
-#TODO version classifiers may not be needed.
+# TODO figure out how to replace this
+# packages = [{ include = "cynetdiff", from = "src" }]
+
 classifiers = [
     "Development Status :: 3 - Alpha",
     "License :: OSI Approved :: MIT License",
@@ -24,62 +27,60 @@ classifiers = [
     "Intended Audience :: Science/Research",
 ]
 
-[tool.poetry.urls]
+[project.urls]
 homepage = "https://github.com/eliotwrobson/CyNetDiff"
 tracker = "https://github.com/eliotwrobson/CyNetDiff/issues"
 documentation = "https://eliotwrobson.github.io/CyNetDiff/"
 
-[tool.poetry.dependencies]
-python = ">=3.9"
 
+[tool.hatch.envs.benchmark]
+dependencies = [
+    "ndlib~=5.1.1",
+    "networkx~=3.2.1",
+    "pandas>=2.0.0",
+    "tqdm~=4.66.2",
+    "pooch~=1.8.1",
+    "notebook~=7.1.2",
+    "coloraide~=3.1.2",
+]
 
-[tool.poetry.group.benchmark]
-optional = true
+[tool.hatch.envs.dev]
+dependencies = [
+    "pytest~=8.3.4",
+    "mypy~=1.7.1",
+    "types-setuptools~=69.0.0.0",
+    "networkx~=3.2.1",
+    "numpy~=2.0.0",
+    "ruff~=0.8.4",
+    "pandas-stubs~=2.2.0.240218",
+    "types-tqdm~=4.66.0.20240106",
+    "cython~=3.0.9",
+    "cython-lint~=0.16.2",
+    "pytest-benchmark~=5.1.0",
+]
 
-[tool.poetry.group.benchmark.dependencies]
-python = ">=3.10"
-ndlib = "^5.1.1"
-networkx = "^3.2.1"
-pandas = ">=2.0.0"
-tqdm = "^4.66.2"
-pooch = "^1.8.1"
-notebook = "^7.1.2"
-coloraide = "^3.1.2"
-
-[tool.poetry.group.dev.dependencies]
-pytest = "^8.3.4"
-mypy = "^1.7.1"
-types-setuptools = "^69.0.0.0"
-networkx = "^3.2.1"
-numpy = "^2.0.0"
-ruff = "^0.8.4"
-pandas-stubs = "^2.2.0.240218"
-types-tqdm = "^4.66.0.20240106"
-cython = "^3.0.9"
-cython-lint = "^0.16.2"
-pytest-benchmark = "^5.1.0"
-
-
-[tool.poetry.group.docs]
-optional = true
-
-[tool.poetry.group.docs.dependencies]
-mkdocs = "^1.4.2"
[…]
