+++ .github/workflows/tests.yml
@@ -19,16 +19,14 @@ jobs:
             3.11
             3.12
             3.13
-      - name: Set up pip cache
-        if: runner.os == 'Linux'
-        uses: actions/cache@v3
-        with:
-          path: ~/.cache/pip
-          key: ${{ runner.os }}-pip-${{ hashFiles('pyproject.toml') }}
-          restore-keys: ${{ runner.os }}-pip-
+      # - name: Set up pip cache
+      #   if: runner.os == 'Linux'
+      #   uses: actions/cache@v3
+      #   with:
+      #     path: ~/.cache/pip
+      #     key: ${{ runner.os }}-pip-${{ hashFiles('pyproject.toml') }}
+      #     restore-keys: ${{ runner.os }}-pip-
       - name: Install Hatch
         run: pipx install hatch
-      - name: Build project
-        run: hatch build
       - name: Run tests
         run: hatch test
