### 2025-01-06 eliotwrobson/CyNetDiff#27 @eliotwrobson
@Kai-Striega thanks for the pointers so far! I've started work on this in #34, I think I have the hang of Meson (it's been pretty straightforward to learn), and I think I have the project building using hatch along with the hatch-meson plugin. It looks like the project is building correctly, but when I go to run the test cases in hatch, it looks like the compiled extensions are missing. Do you know of any projects using a similar setup, or whether I have something misconfigured in meson?
