+++ pyproject.toml
@@ -1,9 +1,9 @@
 [project]
 name = "cynetdiff"
-version = "0.1.16" # TODO switch to dynamic versioning
+version = "0.1.16"                                                           # TODO switch to dynamic versioning
 description = "A fast network diffusion library"
 authors = [{ name = "Eliot W. Robson", email = "eliot.robson24@gmail.com" }]
-license = "MIT"
+#license = "MIT"
 requires-python = ">=3.9"
 readme = "README.md"
 maintainers = [
@@ -79,10 +79,24 @@ dependencies = [
 
 [build-system]
 build-backend = "hatchling.build"
-requires = ["numpy", "cython", "ninja", "hatchling", "hatch-meson"]
+requires = ["numpy", "cython", "hatchling", "hatch-cython"]
 
-[tool.hatch.build.hooks.meson]
-[tool.hatch.metadata.hooks.meson]
+[tool.hatch.build.targets.wheel.hooks.cython]
+dependencies = ["hatch-cython"]
+
+[tool.hatch.build.targets.wheel.hooks.cython.options]
+include_numpy = true
+compile_py = false
+define_macros = [["NPY_NO_DEPRECATED_API", "NPY_1_7_API_VERSION"]]
+include_npyrandom = { pkg = "numpy", include = "get_include", libraries = "get_libraries", library_dirs = "get_library_dirs", required_call = "create_library_symlinks" }
+
+[tool.hatch.build.hooks.cython.options]
+src = "cynetdiff"
+directives = { language_level = 3 }
+
+[tool.hatch.build.targets.wheel]
+packages = ["cynetdiff"]
+
 
 ### Start of static analyis tools ###
 
