### 2025-01-07 eliotwrobson/CyNetDiff#27 @Kai-Striega
I don't know any project using hatch :(

I'm pretty busy this week with work, I'll try to take a more thorough look at this once I have time. Hopefully within the next fortnight.
