### 2025-01-09 0552d7d
Changed lint config
