+++ .github/workflows/lint.yml
@@ -25,20 +25,21 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
 
-      - uses: abatilo/actions-poetry@v2
-        with:
-          poetry-version: ${{ matrix.poetry-version }}
+      - name: Install PDM
+        run: |
+          python -m pip install pdm
+
       - name: Install dependencies
-        run: poetry install --with dev,benchmark
+        run: pdm install
 
       - name: Run ruff lint check
-        run: poetry run ruff check .
+        run: pdm run ruff check .
 
       - name: Run ruff format check
-        run: poetry run ruff format --check .
+        run: pdm run ruff format --check .
 
       - name: Run Cython lint
-        run: poetry run cython-lint .
+        run: pdm run cython-lint .
 
       - name: Run mypy
-        run: poetry run mypy .
+        run: pdm run mypy .
