+++ .github/workflows/lint.yml
@@ -26,11 +26,10 @@ jobs:
           python-version: ${{ matrix.python-version }}
 
       - name: Install PDM
-        run: |
-          python -m pip install pdm
+        run: python -m pip install pdm
 
       - name: Install dependencies
-        run: pdm install
+        run: pdm install -G lint
 
       - name: Run ruff lint check
         run: pdm run ruff check .
