+++ .github/workflows/lint.yml
@@ -26,10 +26,10 @@ jobs:
           python-version: ${{ matrix.python-version }}
 
       - name: Install PDM
-        run: python -m pip install pdm
+        run: pipx install pdm
 
       - name: Install dependencies
-        run: pdm install -G lint
+        run: pdm install -G lint,build-deps
 
       - name: Run ruff lint check
         run: pdm run ruff check .
