+++ .github/workflows/lint.yml
@@ -29,7 +29,7 @@ jobs:
         run: pipx install pdm
 
       - name: Install dependencies
-        run: pdm install -G lint,build-deps
+        run: pdm install -G lint,build
 
       - name: Run ruff lint check
         run: pdm run ruff check .
