+++ pyproject.toml
@@ -29,7 +29,7 @@ tracker = "https://github.com/eliotwrobson/CyNetDiff/issues"
 documentation = "https://eliotwrobson.github.io/CyNetDiff/"
 
 [dependency-groups]
-build-deps = [
+build = [
     "setuptools",
     "meson-python>=0.16.0",
     "ninja>=1.8.2",
