+++ pyproject.toml
@@ -51,7 +51,8 @@ lint = [
     "pandas-stubs~=2.2.0.240218",
     "types-tqdm~=4.66.0.20240106",
     "cython-lint~=0.16.2",
-    "pytest",
+    "pytest~=8.3.4",
+    "coloraide~=3.1.2",
 ]
 docs = [
     "mkdocs~=1.4.2",
