+++ .github/workflows/lint.yml
@@ -25,8 +25,8 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
 
-      - name: Install PDM
-        run: pipx install pdm
+      - name: Setup PDM
+        uses: pdm-project/setup-pdm@v4
 
       - name: Install dependencies
         run: pdm install -G lint,build
