+++ .github/workflows/tests.yml
@@ -25,7 +25,7 @@ jobs:
           cache: true
 
       - name: Install dependencies
-        run: pdm install -G test
+        run: pdm install -G test,build
 
       - name: Run tests
         run: pdm run pytest
