+++ .github/workflows/tests.yml
@@ -7,7 +7,7 @@ on:
     branches: ["*"]
 
 jobs:
-  ci:
+  test:
     strategy:
       matrix:
         python-version: ["3.9", "3.10", "3.11", "3.12", "3.13"]
@@ -29,9 +29,3 @@ jobs:
 
       - name: Run tests
         run: pdm run pytest
-
-      #- name: Install dependencies
-      #  run: pdm install -G test
-
-      #- name: Run tests
-      #  run: pdm run pytest
