+++ .github/workflows/lint.yml
@@ -10,23 +10,23 @@ on:
     branches: ["*"]
 
 jobs:
-  ci:
+  lint:
     strategy:
       matrix:
         python-version: ["3.12"]
-        poetry-version: ["1.8.1"]
         os: [ubuntu-latest]
     runs-on: ${{ matrix.os }}
     steps:
       - uses: actions/checkout@v4
-        with:
-          submodules: recursive
       - uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
 
       - name: Setup PDM
         uses: pdm-project/setup-pdm@v4
+        with:
+          python-version: ${{ matrix.python-version }}
+          cache: true
 
       - name: Install dependencies
         run: pdm install -G lint,build
