+++ .github/workflows/publish.yml
@@ -5,7 +5,7 @@ on:
     tags:
       - "v*.*.*"
 jobs:
-  build:
+  publish:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
