+++ .github/workflows/benchmark.yml
@@ -8,26 +8,24 @@ on:
     branches: ["*"]
 
 jobs:
-  build:
+  benchmark:
+    strategy:
+      matrix:
+        python-version: ["3.12"]
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v5
         with:
-          python-version: |
-            3.9
-            3.10
-            3.11
-            3.12
-            3.13
-      - name: Set up pip cache
-        if: runner.os == 'Linux'
-        uses: actions/cache@v3
+          python-version: ${{ matrix.python-version }}
+      - name: Setup PDM
+        uses: pdm-project/setup-pdm@v4
         with:
-          path: ~/.cache/pip
-          key: ${{ runner.os }}-pip-${{ hashFiles('pyproject.toml') }}
-          restore-keys: ${{ runner.os }}-pip-
-      - name: Install Hatch
-        run: pipx install hatch
+          python-version: ${{ matrix.python-version }}
+          cache: true
+
+      - name: Install dependencies
+        run: pdm install -G test,build
+
       - name: Run benchmarks
-        run: hatch test benchmark
+        run: pdm run pytest benchmarks
