### 2025-01-09 71e5948
simplified a couple things
