+++ .github/workflows/lint.yml
@@ -11,25 +11,21 @@ on:
 
 jobs:
   lint:
-    strategy:
-      matrix:
-        python-version: ["3.12"]
-        os: [ubuntu-latest]
-    runs-on: ${{ matrix.os }}
+    runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v5
         with:
-          python-version: ${{ matrix.python-version }}
+          python-version: 3.12
 
       - name: Setup PDM
         uses: pdm-project/setup-pdm@v4
         with:
-          python-version: ${{ matrix.python-version }}
+          python-version: 3.12
           cache: true
 
       - name: Install dependencies
-        run: pdm install -G lint,build
+        run: pdm install -G lint --no-build
 
       - name: Run ruff lint check
         run: pdm run ruff check .
