+++ .github/workflows/docs.yml
@@ -18,9 +18,8 @@ jobs:
       - uses: actions/setup-python@v5
         with:
           python-version: 3.12
-      - uses: abatilo/actions-poetry@v2
-        with:
-          poetry-version: 1.8.2
+      - name: Setup PDM
+        uses: pdm-project/setup-pdm@v4
       - run: echo "cache_id=$(date --utc '+%V')" >> $GITHUB_ENV
       - uses: actions/cache@v3
         with:
@@ -28,5 +27,5 @@ jobs:
           path: .cache
           restore-keys: |
             mkdocs-material-
-      - run: poetry install --with docs
-      - run: poetry run mkdocs gh-deploy --force
+      - run: pdm install -G docs --no-build
+      - run: pdm run mkdocs gh-deploy --force
