+++ .github/workflows/lint.yml
@@ -25,7 +25,7 @@ jobs:
           cache: true
 
       - name: Install dependencies
-        run: pdm install -G lint --no-build
+        run: pdm install -G lint --no-self
 
       - name: Run ruff lint check
         run: pdm run ruff check .
