+++ .github/workflows/tests.yml
@@ -15,9 +15,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     steps:
       - uses: actions/checkout@v4
-      - uses: actions/setup-python@v5
-        with:
-          python-version: ${{ matrix.python-version }}
+
       - name: Setup PDM
         uses: pdm-project/setup-pdm@v4
         with:
