+++ pyproject.toml
@@ -35,6 +35,7 @@ build = [
     "ninja>=1.8.2",
     "cython",
     "numpy~=2.0.0",
+    "cibuildwheel>=2.0.0",
 ]
 test = [
     "pytest",
@@ -68,7 +69,11 @@ install = ["--no-isolation"]
 
 [build-system]
 build-backend = "mesonpy"
-requires = ["meson-python>=0.16.0", "cython", "numpy"]
+requires = ["meson-python>=0.16.0", "cython", "numpy~=2.0.0", "ninja>=1.8.2"]
+
+[tool.cibuildwheel]
+build = "cp38-* cp39-* cp310-* cp311-*"
+test-command = "pytest {project}/tests"
 
 # [tool.hatch.envs.hatch-test]
 # extra-dependencies = [
