+++ .github/workflows/lint.yml
@@ -14,9 +14,6 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
-      - uses: actions/setup-python@v5
-        with:
-          python-version: 3.12
 
       - name: Setup PDM
         uses: pdm-project/setup-pdm@v4
