### 2025-01-09 ea0be08
Added build frontend
