### 2025-01-09 eliotwrobson/CyNetDiff#27 @eliotwrobson
A quick update: In the current PR #34, I've managed to get the tests and linting working with PDM, but I'm getting the following error message when I try to run `cibuildwheel`:

![image](https://github.com/user-attachments/assets/a3d4db20-a4e0-4db3-8167-b5909139aff9)

It looks like the compiled extension isn't getting installed for this test for some reason? These same tests work when I run `pdm install`, and it looks like the compiled file is present in the wheel files themselves. @Kai-Striega have you ever seen anything like this before?
