### 2025-01-21 eliotwrobson/CyNetDiff#27 @eli-schwartz
> It looks like the compiled extension isn't getting installed for this test for some reason? These same tests work when I run `pdm install`, and it looks like the compiled file is present in the wheel files themselves. [@Kai-Striega](https://github.com/Kai-Striega) have you ever seen anything like this before?

It actually shouldn't matter whether the extension is getting installed. cibuildwheel intentionally runs the tests via `test-comand = "pytest {project}/tests"` as documented:

> Shell command to run tests after the build. The wheel will be installed automatically and available for import from the tests. To ensure the wheel is imported by your tests (instead of your source copy), tests are not run from your project directory. Use the placeholders {project} and {package} when specifying paths in your project. If this variable is not set, your wheel will not be installed after building.

So, `/project/cynetdiff/utils.py` shouldn't be on python's `sys.path` as `from cynetdiff.utils import ......` at all. The fact that it is, implies something else went wrong.

If it was just the wheel not being installed, I would assume that it couldn't import cynetdiff at all -- not that it would import an incorrect version.
