### 2025-01-23 eliotwrobson/CyNetDiff#27 @eliotwrobson
> So, `/project/cynetdiff/utils.py` shouldn't be on python's `sys.path` as `from cynetdiff.utils import ......` at all. The fact that it is, implies something else went wrong.
>
> If it was just the wheel not being installed, I would assume that it couldn't import cynetdiff at all -- not that it would import an incorrect version.

@eli-schwartz thanks for your comment! how would you expect to be able to import the utils file given the current configuration? I should be able to play around with this some more in a few days and try debugging some more.
