### 2025-04-01 #175 @eliotwrobson
@Kai-Striega thanks for checking in! I don't think we need anything else from your end since your review was already completed, and all the items besides adding additional wheels have been completed.

The slowness here has just been from my side, I spent a few days trying to get additional wheels working to no avail, and I've been struggling to find more time to take another crack at this. I think there might be something wrong with the project configuration (PR is here https://github.com/eliotwrobson/CyNetDiff/pull/34), and this is causing the built wheels to fail testing from `cibuildwheel`. However, I'm still not sure how this is configured incorrectly. I finally have some more free time, so I can take another crack at this in a few days.

I recognize this change increases the usability of the package, but it might be worth closing out this review but leaving this as an open issue.
