+++ .github/workflows/docs.yml
@@ -15,12 +15,12 @@ jobs:
         run: |
           git config user.name github-actions[bot]
           git config user.email 41898282+github-actions[bot]@users.noreply.github.com
+      # TODO get rid of python version here once you get this working.
       - uses: actions/setup-python@v5
         with:
           python-version: 3.12
-      - uses: abatilo/actions-poetry@v2
-        with:
-          poetry-version: 1.8.2
+      - name: Setup PDM
+        uses: pdm-project/setup-pdm@v4
       - run: echo "cache_id=$(date --utc '+%V')" >> $GITHUB_ENV
       - uses: actions/cache@v3
         with:
@@ -28,5 +28,5 @@ jobs:
           path: .cache
           restore-keys: |
             mkdocs-material-
-      - run: poetry install --with docs
-      - run: poetry run mkdocs gh-deploy --force
+      - run: pdm install -G docs --no-self
+      - run: pdm run mkdocs gh-deploy --force
