+++ pyproject.toml
@@ -4,8 +4,8 @@ version = "0.1.18" # TODO switch to dynamic versioning
 description = "A fast network diffusion library"
 authors = [{ name = "Eliot W. Robson", email = "eliot.robson24@gmail.com" }]
 license = "MIT"
-license-files = ["LICENSE.txt"]
-requires-python = ">=3.9"
+license-files = ["LICENSE"]
+requires-python = ">=3.10"
 readme = "README.md"
 maintainers = [
     { name = "Eliot W. Robson", email = "eliot.robson24@gmail.com" },
@@ -14,7 +14,6 @@ maintainers = [
 
 classifiers = [
     "Development Status :: 3 - Alpha",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
@@ -31,11 +30,13 @@ documentation = "https://eliotwrobson.github.io/CyNetDiff/"
 [dependency-groups]
 build = [
     "setuptools",
-    "meson-python>=0.16.0",
+    "meson-python>=0.18.0",
     "ninja>=1.8.2",
     "cython",
     "numpy~=2.0.0",
     "cibuildwheel>=2.0.0",
+    "pythran",
+    "pybind11",
 ]
 test = [
     "pytest",
@@ -69,7 +70,7 @@ install = ["--no-isolation"]
 
 [build-system]
 build-backend = "mesonpy"
-requires = ["meson-python>=0.16.0", "cython", "numpy~=2.0.0", "ninja>=1.8.2"]
+requires = ["meson-python>=0.18.0", "cython", "numpy~=2.0.0", "ninja>=1.8.2"]
 
 [tool.cibuildwheel]
 build = "cp38-* cp39-* cp310-* cp311-*"
