+++ pyproject.toml
@@ -49,7 +49,7 @@ lint = [
     "mypy~=1.7.1",
     "networkx~=3.2.1",
     "numpy~=2.0.0",
-    "ruff~=0.8.4",
+    "ruff~=0.11.13",
     "pandas-stubs~=2.2.0.240218",
     "types-tqdm~=4.66.0.20240106",
     "cython-lint~=0.16.2",
