+++ .github/workflows/publish.yml
@@ -13,7 +13,7 @@ jobs:
   publish:
     strategy:
       matrix:
-        python-version: ["3.9"] #, "3.10", "3.11", "3.12", "3.13"]
+        python-version: ["3.10"] #, "3.10", "3.11", "3.12", "3.13"]
         os: [ubuntu-latest]
     runs-on: ${{ matrix.os }}
     steps:
