### 2025-06-15 d8f5346
Simplified publishing
