+++ pyproject.toml
@@ -73,8 +73,7 @@ build-backend = "mesonpy"
 requires = ["meson-python>=0.18.0", "cython", "numpy~=2.0.0", "ninja>=1.8.2"]
 
 [tool.cibuildwheel]
-build = "cp38-* cp39-* cp310-* cp311-*"
-build-frontend = "build"
+build-frontend = "build[uv]"
 test-requires = "pytest networkx numpy"
 test-command = "pytest {project}/tests"
 
