+++ .github/workflows/lint.yml
@@ -18,7 +18,7 @@ jobs:
       - name: Setup PDM
         uses: pdm-project/setup-pdm@v4
         with:
-          python-version: 3.12
+          python-version: 3.10
           cache: true
 
       - name: Install dependencies
