+++ .github/workflows/tests.yml
@@ -15,6 +15,7 @@ jobs:
       - name: Setup PDM
         uses: pdm-project/setup-pdm@v4
         with:
+          version: latest
           python-version: 3.10
           cache: true
 
