+++ .github/workflows/benchmark.yml
@@ -9,9 +9,6 @@ on:
 
 jobs:
   benchmark:
-    strategy:
-      matrix:
-        python-version: ["3.12"]
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
@@ -19,7 +16,8 @@ jobs:
       - name: Setup PDM
         uses: pdm-project/setup-pdm@v4
         with:
-          python-version: ${{ matrix.python-version }}
+          version: latest
+          python-version: 3.12
           cache: true
 
       - name: Install dependencies
