+++ .github/workflows/tests.yml
@@ -17,7 +17,6 @@ jobs:
         with:
           version: latest
           python-version: 3.10
-          cache: true
 
       - name: Install dependencies
         run: pdm install -G test,build
