+++ .github/workflows/publish.yml
@@ -3,9 +3,9 @@
 name: publish
 on:
   workflow_dispatch:
-  release:
-    types:
-      - published
+    release:
+      types:
+        - published
     pull_request:
       paths:
         - .github/workflows/publish.yml
