+++ .github/workflows/lint.yml
@@ -20,7 +20,6 @@ jobs:
         with:
           version: latest
           python-version: 3.10
-          cache: true
 
       - name: Install dependencies
         run: pdm install -G lint --no-self
