+++ .github/workflows/tests.yml
@@ -16,7 +16,8 @@ jobs:
         uses: pdm-project/setup-pdm@v4
         with:
           version: latest
-          python-version: 3.10
+          python-version: "3.10"
+          cache: true
 
       - name: Install dependencies
         run: pdm install -G test,build
