+++ .github/workflows/benchmark.yml
@@ -17,7 +17,7 @@ jobs:
         uses: pdm-project/setup-pdm@v4
         with:
           version: latest
-          python-version: 3.12
+          python-version: "3.12"
 
       - name: Install dependencies
         run: pdm install -G test,build
