+++ .github/workflows/lint.yml
@@ -19,7 +19,8 @@ jobs:
         uses: pdm-project/setup-pdm@v4
         with:
           version: latest
-          python-version: 3.10
+          python-version: "3.10"
+          cache: true
 
       - name: Install dependencies
         run: pdm install -G lint --no-self
