+++ .github/workflows/benchmark.yml
@@ -16,8 +16,8 @@ jobs:
       - name: Setup PDM
         uses: pdm-project/setup-pdm@v4
         with:
-          version: latest
           python-version: "3.12"
+          cache: true
 
       - name: Install dependencies
         run: pdm install -G test,build
