+++ .github/workflows/publish.yml
@@ -36,6 +36,8 @@ jobs:
     steps:
       - uses: actions/checkout@v4
 
+      - uses: actions/setup-python@v5
+
       - uses: astral-sh/setup-uv@v6
 
       - uses: pypa/cibuildwheel@v2.23
