+++ .github/workflows/publish.yml
@@ -37,6 +37,8 @@ jobs:
       - uses: actions/checkout@v4
 
       - uses: actions/setup-python@v5
+        with:
+          python-version: "3.11"
 
       - uses: astral-sh/setup-uv@v6
 
