+++ pyproject.toml
@@ -73,6 +73,7 @@ build-backend = "mesonpy"
 requires = ["meson-python>=0.18.0", "cython", "numpy~=2.0.0", "ninja>=1.8.2"]
 
 [tool.cibuildwheel]
+build = "cp310-* cp311-* cp312-* cp313-*"
 build-frontend = "build[uv]"
 test-requires = "pytest networkx numpy"
 test-command = "pytest {project}/tests"
