+++ .github/workflows/docs.yml
@@ -15,10 +15,7 @@ jobs:
         run: |
           git config user.name github-actions[bot]
           git config user.email 41898282+github-actions[bot]@users.noreply.github.com
-      # TODO get rid of python version here once you get this working.
       - uses: actions/setup-python@v5
-        with:
-          python-version: 3.12
       - name: Setup PDM
         uses: pdm-project/setup-pdm@v4
       - run: echo "cache_id=$(date --utc '+%V')" >> $GITHUB_ENV
