+++ .github/workflows/publish.yml
@@ -37,10 +37,6 @@ jobs:
     steps:
       - uses: actions/checkout@v4
 
-      - uses: actions/setup-python@v5
-        with:
-          python-version: "3.11"
-
       - uses: astral-sh/setup-uv@v6
 
       - uses: pypa/cibuildwheel@v2.23
