+++ tests/test_ic_model.py
@@ -198,7 +198,7 @@ def test_randomized_activation(seed: int) -> None:
 
     activated_nodes = set(model.get_activated_nodes())
 
-    for seed_node, prob in zip(seeds, seed_probs):
+    for seed_node, prob in zip(seeds, seed_probs, strict=True):
         assert (seed_node in activated_nodes) == (prob == 1.0)
 
 
@@ -445,7 +445,7 @@ def test_marginal_gain(directed: bool, include_payoffs: bool, seed: int) -> None
 
     set_so_far: t.List[int] = []
 
-    for seed, result in zip(seeds, results[1:]):
+    for seed, result in zip(seeds, results[1:], strict=False):
         without_new_seed_total = compute_graph_marginal_gain(test_graph, independent_cascade(test_graph, set_so_far))
 
         with_new_seed_total = compute_graph_marginal_gain(
