+++ pyproject.toml
@@ -57,11 +57,11 @@ lint = [
     "coloraide~=3.1.2",
 ]
 docs = [
-    "mkdocs~=1.4.2",
-    "mkdocs-material~=9.1.9",
-    "mkdocs-macros-plugin~=0.7.0",
-    "mkdocstrings~=0.23.0",
-    "mkdocstrings-python~=1.7.3",
+    "mkdocs~=1.6.1",
+    "mkdocs-material~=9.6.14",
+    "mkdocs-macros-plugin~=1.3.7",
+    "mkdocstrings~=0.29.1",
+    "mkdocstrings-python~=1.16.12",
     "setuptools~=69.0.3",
 ]
 
