### 2025-06-16 eliotwrobson/CyNetDiff#27 @eliotwrobson
Resolved with https://github.com/eliotwrobson/CyNetDiff/pull/36
