### 2025-06-16 #175 @eliotwrobson
Apologies all for the very delayed action on my part (wrapping up my PhD next month and things have been extremely busy). With the [latest release](https://github.com/eliotwrobson/CyNetDiff/releases/tag/v0.1.18) I was _finally_ able to provide wheels for common platforms! I wasn't able to include windows because of some issues with the build workflow but mac and linux are covered (see the PR [here](https://github.com/eliotwrobson/CyNetDiff/pull/36)). I believe this just means windows users need to compile locally, but I can see about fixing this if users run into issues later.

@sneakers-the-rat is the remaining step to get final sign off from the reviewers? I think from the discussion the only outstanding item was having the wheels built, so not sure if the protocol is to wait for their response.
