### 2025-08-04 #175 @lwasser
Hi friends. I'm checking in on this review again, given our EIC is now the project author!
@Kai-Striega @jeremyzhangsq from your end, has @eliotwrobson addressed your review comments to your satisfaction?
