### 2025-08-16 #175 @Kai-Striega
This looks great, all my previous comments have been addressed, just a couple of small comments:

- [x] [CONTRIBUTING.md](https://github.com/eliotwrobson/CyNetDiff/blob/main/CONTRIBUTING.md) is outdated - it is still saying the project uses poetry (which I'm pretty sure it now doesn't
- [x] I can't find any documentation for how I _should_ build the project using pdm/meson

@lwasser @eliotwrobson
