### 2025-08-17 #175 @eliotwrobson
@Kai-Striega thanks for following up and catching this! I've updated the contributing file and added directions about how to build the project with pdm locally: https://github.com/eliotwrobson/CyNetDiff/blob/main/CONTRIBUTING.md
