### 2025-08-22 #175 @lwasser
Ok everyone!! Thank you for your work on this review.

---

🎉 CyNetDiff has been approved by pyOpenSci! Thank you, @eliotwrobson , for submitting the package, and many thanks to @Kai-Striega and @jeremyzhangsq  for reviewing this package! 😸

- [x] Add the badge for pyOpenSci peer-review to the README.md of <package-name-here>. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

- [ ] Eliot, are you interested in writing a blog post about CyNetDiff? Since I know the packaging part was tricky, people might also be interested in learning about your experiences.
- [x] Eliot, you are already in Slack, so this step is done!

---
