### 2025-08-22 #175 @lwasser
@eliotwrobson when you have a final release and a Zenodo DOI, please share it here so I can update the header (or drop it in there yourself!)
