### 2025-08-23 #175 @eliotwrobson
@lwasser the zenodo DOI is [10.5281/zenodo.10801503](https://doi.org/10.5281/zenodo.10801503). All of the above tasks have been completed besides the other maintainer @abhishekumrawal completing the survey! I did all of the primary work that went into the release, so I can reach out to him and I think we can wrap this up.

Yes, a blog post would be good. Since I used a significant amount of AI assistance for the packaging, I think we should probably wait until our AI policy is finalized, then the post can highlight this policy as part of the discussion. What are your thoughts?
